% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.cache.R
\name{clear.cache}
\alias{clear.cache}
\title{Clear MatchingPursuit Cache}
\usage{
clear.cache()
}
\value{
Logical scalar. TRUE if all files were successfully removed, FALSE otherwise.
The return value is invisible.
}
\description{
Deletes all files in the MatchingPursuit cache directory.
}
\examples{
if (interactive()) {
  clear.cache()
}
}
