% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear MatchingPursuit cache}
\usage{
clear_cache()
}
\value{
Logical scalar. Returns \code{TRUE} if all files were successfully removed,
and \code{FALSE} otherwise. The return value is invisible.
}
\description{
Deletes all files in the MatchingPursuit cache directory.
}
\examples{
if (interactive()) {
  clear_cache()
}
}
