% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf-methods.R
\name{edf-methods}
\alias{edf-methods}
\alias{plot.edf}
\alias{print.edf}
\alias{summary.edf}
\alias{print.summary.edf}
\title{Methods for EDF Objects}
\usage{
\method{plot}{edf}(
  x,
  begin = NULL,
  end = NULL,
  panel_height = NULL,
  rainbow = FALSE,
  bg_colour = "white",
  txt_col = "black",
  zero_line = TRUE,
  main = NULL,
  ...
)

\method{print}{edf}(x, ...)

\method{summary}{edf}(object, ...)

\method{print}{summary.edf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"edf"} or \code{"summary.edf"}.}

\item{begin}{Time point (in seconds) at which to start plotting.
If \code{NULL}, plotting starts at the beginning of the signal.}

\item{end}{Time point (in seconds) at which to stop plotting.
If \code{NULL}, plotting continues to the end of the signal.}

\item{panel_height}{Controls the vertical spacing between individual signals.
If \code{NULL}, the value is chosen automatically.}

\item{rainbow}{Logical. If \code{TRUE}, individual channels are drawn
in different colours.}

\item{bg_colour}{Background colour.}

\item{txt_col}{Colour of text elements.}

\item{zero_line}{Logical. If \code{TRUE}, a horizontal line representing
zero is displayed for each channel.}

\item{main}{Plot title. If \code{NULL}, the EDF record name is used.}

\item{...}{Additional arguments. Currently ignored.}

\item{object}{An object of class \code{"edf"}.}
}
\value{
\code{print.edf()} and \code{print.summary.edf()} return their input
object invisibly.

\code{summary.edf()} returns an object of class \code{"summary.edf"}
containing record information, signal dimensions, sampling information,
channel names, and basic descriptive statistics for each channel.

\code{plot.edf()} is called for its side effect and returns no value.
}
\description{
Methods for printing, summarizing, and plotting objects of class
\code{"edf"}, together with a print method for objects returned by
\code{summary()}.
}
\details{
Objects of class \code{"edf"} represent multi-channel signals imported
from EDF files, together with sampling information, channel names,
a time vector, and the source record name.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
x <- read_edf_signals(file, resampling = FALSE)

print(x)
summary(x)

plot(
  x,
  begin = 0,
  end = 10,
  panel_height = NULL,
  rainbow = TRUE,
  bg_colour = "black",
  txt_col = "white",
  zero_line = TRUE,
  main = "EEG signals stored in the EEG.edf file"
)

plot(
  x,
  begin = 0,
  end = 10,
  panel_height = NULL,
  rainbow = FALSE,
  bg_colour = "white",
  txt_col = "black",
  zero_line = TRUE,
  main = "EEG signals stored in the EEG.edf file"
)
}
\seealso{
\code{\link{read_edf_signals}}
}
