% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg_montage.R
\name{eeg_montage}
\alias{eeg_montage}
\title{Performs bipolar, reference or average EEG montage}
\usage{
eeg_montage(
  x,
  montage_type = c("average", "reference", "bipolar"),
  ref_channel = NULL,
  bipolar_pairs = NULL
)
}
\arguments{
\item{x}{Object of class \code{edf} (from \code{read_edf_signals()}) or
data frame with samples in rows and channels in columns.
The data frame must have column names corresponding to channel names.}

\item{montage_type}{A character string specifying the montage type.
\itemize{
   \item \code{"average"} - each electrode is referenced to the average of all electrodes
   \item \code{"reference"} - each active electrode is compared to a single common reference electrode
   \item \code{"bipolar"} - each channel compares two adjacent electrodes
}}

\item{ref_channel}{Name of the reference channel for \code{"reference"} montage.}

\item{bipolar_pairs}{List of electrodes pairs for \code{"bipolar"} montage. See example below.}
}
\value{
An object of class \code{edf}, which is a list with fields:

\item{signal}{Data frame containing all signal channels.}
\item{sampling_frequency}{Sampling frequency.}
\item{time}{Time stamps.}
\item{signal_names}{Names of the signal channels.}
\item{record_name}{Name of the EDF file.}
}
\description{
An EEG montage refers to the arrangement of EEG electrodes and the way their signals
are displayed relative to one another during electroencephalogram interpretation.
The same EEG recording may appear very different depending on the montage used. This
function implements the three montage methods most commonly used in practice:
1) Bipolar Montage, 2) Referential (Monopolar) Montage, and
3) Average Reference Montage.
}
\details{
To check the channel names in the analysed EEG recording,
use the \code{read_edf_params()} function.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out <- read_edf_signals(file, resampling = FALSE, from = 0, to = 10)

read_edf_params(file)

# The classical double banana montage.
pairs <- list(
  c("Fp2", "F4"),
  c("F4",  "C4"),
  c("C4",  "P4"),
  c("P4",  "O2"),
  c("Fp1", "F3"),
  c("F3",  "C3"),
  c("C3",  "P3"),
  c("P3",  "O1"),
  c("Fp2", "F8"),
  c("F8",  "T4"),
  c("T4",  "T6"),
  c("T6",  "O2"),
  c("Fp1", "F7"),
  c("F7",  "T3"),
  c("T3",  "T5"),
  c("T5",  "O1"),
  c("Fz",  "Cz"),
  c("Cz",  "Pz")
)

signal_bip_mont <- eeg_montage(out, montage_type = "bipolar", bipolar_pairs = pairs)
signal_ref_mont <- eeg_montage(out, montage_type = "reference", ref_channel = "O1")
signal_avg_mont <- eeg_montage(out, montage_type = "average")

head(signal_bip_mont$signal)
head(signal_ref_mont$signal)
head(signal_avg_mont$signal)

}
