% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.download.R
\name{empi.download}
\alias{empi.download}
\title{Downloads \emph{Enhanced Matching Pursuit Implementation} program \href{https://github.com/develancer/empi}{empi}.}
\usage{
empi.download()
}
\value{
The function downloads the empi program in a version compatible with the operating system used (Windows, Linux, MacOS-x64, MacOS-arm64) and saves it in the current directory.
}
\description{
empi is a fast implementation of Matching Pursuit algorithm. The program uses both CPU parallelization and GPU devices. See README.md for more details on this implementation.
}
\examples{
## Not run:
empi.download()
## End(Not run)

}
