% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.find.R
\name{empi.find}
\alias{empi.find}
\title{Checks if EMPI external software is installed}
\usage{
empi.find()
}
\value{
If the EMPI program is found, it returns its full path. Otherwise, it displays a message
to the user asking them to install it using the \code{empi.install()} function.
}
\description{
The EMPI program is installed using the \code{empi.install()} function and is loaded into the
cache directory. This function checks whether the EMPI program is still there (the user
has free access to the cache directory and can, for example, delete it).
}
\examples{
## Not run:
empi.find()
## End(Not run)

}
