% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.locate.R
\name{empi.locate}
\alias{empi.locate}
\title{Get required external software localization}
\usage{
empi.locate()
}
\value{
List with URL of the EMPI binaries and zip file name.
}
\description{
Returns \strong{Enhanced Matching Pursuit Implementation} binary locations for
the following operation systems: Windows, Linux, MacOS-x64, MacOS-arm64.
}
\examples{
empi.locate()

}
