% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi2tf.R
\name{empi2tf}
\alias{empi2tf}
\title{Creates a time-frequency map using atoms from the Matching Pursuit algorithm}
\usage{
empi2tf(
  db.file = NULL,
  db.list = NULL,
  channel,
  mode = "sqrt",
  freq.divide = 1,
  increase.factor = 1,
  shortening.factor.x = 2,
  shortening.factor.y = 2,
  display.crosses = TRUE,
  display.atom.numbers = FALSE,
  display.grid = FALSE,
  crosses.color = "white",
  palette = "my custom palette",
  rev = TRUE,
  out.mode = "plot",
  path = NULL,
  file.name = NULL,
  size = c(512, 512),
  draw.ellipses = FALSE,
  plot.signals = TRUE,
  write.atoms = FALSE
)
}
\arguments{
\item{db.file}{The SQLite file created after executing the \code{empi.execute()} function.
In this case, the \code{db.list} parameter must be \code{NULL}.}

\item{db.list}{The list created after executing the \code{empi.execute()} function.
In this case, the \code{db.file} parameter must be \code{NULL}.}

\item{channel}{Channel from the SQLite file to process.}

\item{mode}{\code{"sqrt"}, \code{"log"}, or \code{"linear"}. It determines the intensity
with which the so-called blobs are displayed on the T-F map.}

\item{freq.divide}{Specifies how many times the displayed frequency in the T-F map
should be decreased. For example, if the sampling frequency is \code{f=256Hz}, the maximum
frequency in the T-F map will be \code{f/2/freq.divide} (f/2 is the Nyquist rule).}

\item{increase.factor}{Factor of increasing the number of pixels in the f-axis, the most
sensible are non-negative integers (e.g. 2, 4, 5, 8).}

\item{shortening.factor.x}{Usually, for better visualization of atoms, a value of 2 will
be appropriate.}

\item{shortening.factor.y}{Usually, for better visualization of atoms, a value of 2 will
be appropriate.}

\item{display.crosses}{Whether small crosses should be displayed in the canters of atoms.}

\item{display.atom.numbers}{Whether atom numbers should be displayed in the canters of atoms.}

\item{display.grid}{Whether to draw grid lines.}

\item{crosses.color}{Colour of small crosses.}

\item{palette}{Palette from the list returned by \code{hcl.pals()} function or the string
\code{"my custom palette"}.}

\item{rev}{\code{rev} param in \code{hcl.colors()} function.}

\item{out.mode}{One of the following:
\itemize{
   \item \code{"plot"} - draws a T-F map on the screen.
   \item \code{"file"} - saves a T-F map to file \code{file.name} (as \code{png} file).
   \item \code{"RData"} - saves the T-F map of \code{size} in the \code{file.name}
   (as R's matrix), resampling is performed using the function \code{imager::resize()} function.
   \item \code{"RData2"} - saves the T-F map of \code{size} in the \code{file.name}
   (as R's matrix), resampling is performed using the function using \code{raster::resample()} function.
 }}

\item{path}{Path where \code{png}, \code{RData} or \code{pdf} files will be written. If \code{NULL},
files will be written to the cache directory.}

\item{file.name}{Name of the \code{png} file (if \code{out.mode="file}) or name of the \code{RData}
file (if \code{out.mode="RData"} or \code{out.mode="RData2})}

\item{size}{\code{png} file size in pixels (if \code{out.mode="file"}) or size of the T-F matrix
(if \code{out.mode="RData"} of \code{out.mode="RData2"}).}

\item{draw.ellipses}{Only for testing. User can set it to \code{TRUE} to see the effect.
Works properly only if \code{out.mode="plot"}.}

\item{plot.signals}{Whether the original and reconstructed signals should also be displayed.}

\item{write.atoms}{If \code{TRUE}, writes all atom plots into \code{Atoms.pdf} file
(to the cache directory or user specified one - depending on \code{path} variable)}
}
\value{
Depending on the \code{out.mode} parameter the function returns:
   \itemize{
     \item Time-Frequency map plotted on the screen
     \item Time-Frequency map saved in a .png file
     \item Time-Frequency map saved as .RData file
  }
Regardless of the above, the function returns the following:
  \itemize{
    \item all the Gabor functions
    \item reconstructed signal
    \item original signal
    \item sampling frequency
    \item grid size in t axis
    \item grid size in f axis
    \item epoch size in samples
    \item length of the signal in seconds
    \item time-frequency map
    \item time-frequency map after resampling
    (if \code{out.mode="RData"} or if \code{out.mode="RData2"}, otherwise, \code{NULL} is returned)
    \item channel number processed
    \item frequency divide
  }
}
\description{
Creates a time-frequency map using atoms from the Matching Pursuit algorithm.
The created map can be: 1) displayed on the screen, 2) saved in \code{.png} file,
or 3) saved as an \code{.RData} object.
}
\examples{
file <- system.file("extdata", "sample1.db", package = "MatchingPursuit")

out <- empi2tf(
  db.file = file,
  channel = 1,
  mode = "sqrt",
  freq.divide = 4,
  increase.factor= 4,
  display.crosses = TRUE,
  display.atom.numbers = FALSE,
  out.mode = "plot",
)

}
