% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi_install.R
\name{empi_install}
\alias{empi_install}
\title{Installs the EMPI external program}
\usage{
empi_install()
}
\value{
The function downloads the EMPI program in a version compatible
with the operating system used (Windows, Linux, MacOS-x64, MacOS-arm64)
and stores it in the package cache directory.
}
\description{
Downloads the \strong{Enhanced Matching Pursuit Implementation} (EMPI)
external program compatible with the current operating system and stores it
in the package cache directory.
}
\details{
The function detects the operating system (Windows, Linux, macOS arm64),
downloads the appropriate archive from the official
repository, verifies its integrity using a checksum, and extracts it.
}
\examples{
if (interactive()) {
  empi_install()
}

}
