% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi_locate.R
\name{empi_locate}
\alias{empi_locate}
\title{Get required external software localization}
\usage{
empi_locate()
}
\value{
A list containing:
\itemize{
  \item \code{url}: URL of the EMPI binary archive,
  \item \code{fname}: archive file name.
}
}
\description{
Returns \strong{Enhanced Matching Pursuit Implementation} binary locations for
the following operating systems: Windows, Linux, macOS-arm64.
}
\examples{
sys <- Sys.info()[["sysname"]]
mach <- Sys.info()[["machine"]]

if (sys \%in\% c("Windows", "Linux") ||
    (sys == "Darwin" && mach == "arm64")) {
  empi_locate()
}

}
\seealso{
\code{\link{empi_check}},
\code{\link{empi_install}},
\code{\link{empi_execute}},
\code{\link{plot.mp}}
}
