% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.coeff.R
\name{filters.coeff}
\alias{filters.coeff}
\title{A wrapper function for \code{signal::butter()} function}
\usage{
filters.coeff(
  fs = 256,
  notch = c(49, 51),
  notch.order = 2,
  lowpass = 30,
  lowpass.order = 4,
  highpass = 1,
  highpass.order = 4,
  bandpass = c(0.5, 40),
  bandpass.order = 4,
  bandstop = c(0.5, 40),
  bandstop.order = 4
)
}
\arguments{
\item{fs}{Sampling rate.}

\item{notch}{Vector of two frequencies for notch filter.}

\item{notch.order}{Notch filter order.}

\item{lowpass}{Low-pass filter frequency.}

\item{lowpass.order}{Low-pass filter order.}

\item{highpass}{High-pass filter frequency.}

\item{highpass.order}{High-pass filter order.}

\item{bandpass}{Vector of two frequencies for band-pass filter.}

\item{bandpass.order}{Band-pass filter order.}

\item{bandstop}{Vector of two frequencies for band-stop filter.}

\item{bandstop.order}{Band-stop filter order.}
}
\value{
List with parameters of individual filters.
}
\description{
Implements notch, low-pass, high-pass, band-pass, and band-stop filters with desired frequency
ranges and Butterworth filter order.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out <- read.edf.signals(file, resampling = FALSE)
signal <- out$signals
sampling.rate <- out$sampling.rate

fc <- filters.coeff(
  fs = sampling.rate,
  notch = c(49, 51),
  lowpass = 40,
  highpass = 1,
  bandpass = c(0.5, 40),
  bandstop = c(10, 50)
)

print(fc)

signal::freqz(fc$notch, Fs = sampling.rate)
signal::freqz(fc$lowpass, Fs = sampling.rate)
signal::freqz(fc$highpass, Fs = sampling.rate)
signal::freqz(fc$bandpass, Fs = sampling.rate)
signal::freqz(fc$bandstop, Fs = sampling.rate)

plot(signal[, 1], type = "l", panel.first = grid())

signal.filt <- signal

for (m in 1:ncol(signal)) {
  signal.filt[, m] = signal::filtfilt(fc$notch, signal.filt[, m]); # 50Hz notch filter
  signal.filt[, m] = signal::filtfilt(fc$lowpass, signal.filt[, m]); # Low pass IIR Butterworth
  signal.filt[, m] = signal::filtfilt(fc$highpass, signal.filt[, m]); # High pass IIR Butterwoth
}

plot(signal.filt[, 1], type = "l", panel.first = grid())

}
