% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_coeff.R
\name{filters_coeff}
\alias{filters_coeff}
\title{A wrapper function for \code{signal::butter()} function}
\usage{
filters_coeff(
  sampling_frequency = 256,
  notch = c(49, 51),
  notch_order = 2,
  lowpass = 30,
  lowpass_order = 4,
  highpass = 1,
  highpass_order = 4,
  bandpass = c(0.5, 40),
  bandpass_order = 4,
  bandstop = c(0.5, 40),
  bandstop_order = 4
)
}
\arguments{
\item{sampling_frequency}{Sampling frequency.}

\item{notch}{Vector of two frequencies for notch filter.}

\item{notch_order}{Notch filter order.}

\item{lowpass}{Low-pass filter frequency.}

\item{lowpass_order}{Low-pass filter order.}

\item{highpass}{High-pass filter frequency.}

\item{highpass_order}{High-pass filter order.}

\item{bandpass}{Vector of two frequencies for band-pass filter.}

\item{bandpass_order}{Band-pass filter order.}

\item{bandstop}{Vector of two frequencies for band-stop filter.}

\item{bandstop_order}{Band-stop filter order.}
}
\value{
List containing coefficients of individual Butterworth filters.

\describe{
  \item{notch}{Notch filter used to remove a specific narrow frequency band.}
  \item{lowpass}{Low-pass filter that attenuates high-frequency components.}
  \item{highpass}{High-pass filter that attenuates low-frequency components.}
  \item{bandpass}{Band-pass filter that retains frequencies within a selected range.}
  \item{bandstop}{Band-stop filter that removes frequencies within a selected range.}
}
}
\description{
Implements notch, low-pass, high-pass, band-pass, and band-stop filters
with specified frequency ranges and Butterworth filter order.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out <- read_edf_signals(file, resampling = FALSE)
signal <- out$signal
sampling_frequency <- out$sampling_frequency

fc <- filters_coeff(
  sampling_frequency = sampling_frequency,
  notch = c(49, 51),
  lowpass = 40,
  highpass = 1,
  bandpass = c(0.5, 40),
  bandstop = c(10, 50)
)

print(fc)

signal::freqz(fc$notch, Fs =  sampling_frequency)
signal::freqz(fc$lowpass, Fs =  sampling_frequency)
signal::freqz(fc$highpass, Fs =  sampling_frequency)
signal::freqz(fc$bandpass, Fs =  sampling_frequency)
signal::freqz(fc$bandstop, Fs =  sampling_frequency)

plot(signal[, 1], type = "l", panel.first = grid())

signal_filt <- signal

for (m in 1:ncol(signal)) {
  signal_filt[, m] <- signal::filtfilt(fc$notch, signal_filt[, m]); # 50Hz notch filter
  signal_filt[, m] <- signal::filtfilt(fc$lowpass, signal_filt[, m]); # Low pass IIR Butterworth
  signal_filt[, m] <- signal::filtfilt(fc$highpass, signal_filt[, m]); # High pass IIR Butterwoth
}

plot(signal_filt[, 1], type = "l", panel.first = grid())

}
