% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabor_fun.R
\name{gabor_fun}
\alias{gabor_fun}
\title{Gabor function implementation}
\usage{
gabor_fun(
  number_of_samples,
  sampling_frequency,
  mean,
  phase,
  sigma,
  frequency,
  normalization = TRUE
)
}
\arguments{
\item{number_of_samples}{Number of samples in the generated atom.}

\item{sampling_frequency}{Sampling frequency.}

\item{mean}{Time position of the Gaussian envelope.}

\item{phase}{Phase of the sinusoidal component.}

\item{sigma}{Scale parameter controlling the width of the Gaussian window.}

\item{frequency}{Frequency of the sinusoidal component.}

\item{normalization}{If \code{TRUE}, the resulting atom is normalized to have unit norm.}
}
\value{
A list containing four numeric vectors of length \code{number_of_samples}:

  \item{cosine}{Cosine wave.}
  \item{gauss}{Gaussian envelope.}
  \item{gabor}{Gabor function.}
  \item{t}{Time vector corresponding to signal samples.}
}
\description{
A Gabor function is a sinusoidal wave localized by a Gaussian envelope. In signal processing, it is
widely used as a basic building block for representing signals localized in both time and frequency.
The Matching Pursuit algorithm uses a redundant dictionary of so-called \emph{Gabor atoms}.
These atoms are particularly suitable because they: 1) provide optimal time–frequency
localization, 2) represent oscillatory signals well, 3) enable adaptive time-frequency decomposition.
}
\examples{
number_of_samples <- 512
sampling_frequency <- 256.0
mean <- 1
phase <- pi
sigma <- 0.5
frequency <- 5.0
normalization = TRUE

out <- gabor_fun(
  number_of_samples,
  sampling_frequency,
  mean,
  phase,
  sigma,
  frequency,
  normalization
)

# If normalization = TRUE, norm of atom = 1, we can check it
crossprod(out$gabor)

plot(out$t, out$gabor, type = "l", xlab = "t", ylab = "gabor", panel.first = grid())

}
