% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_omp_run_pipeline.R
\name{mp_omp_run_pipeline}
\alias{mp_omp_run_pipeline}
\title{Run the complete MP or OMP decomposition pipeline}
\usage{
mp_omp_run_pipeline(
  mode = NULL,
  sig_file,
  sig_file_type = "csv",
  col_names_in_csv,
  xml_file,
  topk,
  n_nonzero_coefs = NULL,
  tol = NULL,
  normalize = TRUE,
  fit_intercept = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{mode}{\code{"omp"} or \code{"mp"}. Specifies the algorithm to use for signal decomposition.}

\item{sig_file}{Path to a CSV, EDF/EDF(+) or WFDB file containing the signal data.}

\item{sig_file_type}{\code{"csv"} or \code{"wfdb"} or \code{"edf"}. Specifies the input file type.
See \code{read_csv_signals()}, \code{read_wfdb_signals()}, \code{read_edf_signals()} functions.}

\item{col_names_in_csv}{Logical. Indicates whether the CSV file contains column names in the first row.
See \code{read_csv_signals()} function.}

\item{xml_file}{Path to an XML file defining the dictionary atoms.
See \code{read_dict()} function.}

\item{topk}{Integer. Number of atoms with the highest correlation to the
signal retained for MP/OMP fitting.
See \code{topk_atoms()} function.}

\item{n_nonzero_coefs}{Integer. Maximum number of non-zero coefficients  to include in the sparse decomposition.}

\item{tol}{Optional numeric tolerance for the stopping criterion. If \code{NULL}, the
algorithm stops after selecting \code{n_nonzero_coefs} atoms.
If specified, the algorithm stops when the residual energy falls below
this value and overrides n_nonzero_coefs.}

\item{normalize}{Logical. If \code{TRUE}, dictionary atoms are normalized to unit L2 norm before fitting.}

\item{fit_intercept}{Logical. If  \code{TRUE}, an intercept term is included in the model.
Only used when \code{mode == "omp"}.}

\item{verbose}{Logical. If  \code{TRUE}, progress information is printed to  the console.}
}
\value{
An object of class \code{"mp"} containing:

\item{atoms}{A data frame describing the selected atoms.}
\item{original_signal}{Matrix containing the original signal(s).}
\item{reconstruction}{Matrix containing the reconstructed signal(s).}
\item{gabors}{List of matrices containing selected atoms for each channel.}
\item{t}{Time vector corresponding to signal samples.}
\item{sf}{Sampling frequency.}
}
\description{
Executes the complete Orthogonal Matching Pursuit (OMP) or Matching Pursuit
(MP) workflow: (1) reads a signal from a CSV file, (2) loads a dictionary
definition from an XML file, (3) selects the most relevant atoms, and
(4) performs sparse signal decomposition using the selected algorithm.
}
\examples{
sig_file <- system.file("extdata", "sample3.csv", package = "MatchingPursuit")
xml_file <- system.file("extdata", "sample3_dict.xml", package = "MatchingPursuit")

out_mp <- mp_omp_run_pipeline(
  mode = "mp",
  sig_file = sig_file,
  sig_file_type = "csv",
  col_names_in_csv = TRUE,
  xml_file = xml_file,
  topk = 5000,
  n_nonzero_coefs = 50,
  verbose = TRUE
)

out_omp <- mp_omp_run_pipeline(
  mode = "omp",
  sig_file = sig_file,
  sig_file_type = "csv",
  col_names_in_csv = TRUE,
  xml_file = xml_file,
  topk = 5000,
  n_nonzero_coefs = 50,
  verbose = TRUE
)

plot(out_mp, channel = 2)
plot(out_omp, channel = 2)

}
\seealso{
\code{\link{omp_core}},
\code{\link{mp_core}},
\code{\link{mp_omp_execute}},
\code{\link{topk_atoms}},
\code{\link{read_dict}},
}
