% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ecg.R
\name{plot.ecg}
\alias{plot.ecg}
\title{The function displays ECG signals in a layout corresponding to standard paper ECG printouts}
\usage{
\method{plot}{ecg}(
  x,
  begin,
  end,
  panel_height = 3,
  small_squares = TRUE,
  zero_line = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{ecg} (from \code{read_ecg_signals()}).}

\item{begin}{Time point (in seconds) at which to start plotting.}

\item{end}{Time point (in seconds) at which to stop plotting.}

\item{panel_height}{Number of large squares to display (according to standard ECG paper):
\itemize{
    \item small grid: 0.04 sec. x 0.1 mV
    \item large grid: 0.20 sec. x 0.5 mV
}}

\item{small_squares}{If \code{TRUE}, the small grid is also displayed.}

\item{zero_line}{If \code{TRUE}, a horizontal line representing \code{0 mV} is displayed.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize an ECG graph.
}
\description{
A typical ECG paper layout was used, with a small grid of 0.04 s × 0.1 mV and a
large grid of 0.20 s × 0.5 mV.
}
\examples{
# ECG data comes from https://physionet.org/content/ptb-xl/1.0.3/
file <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")
dir <- dirname(file)
name <- tools::file_path_sans_ext(basename(file))

out <- read_ecg_signals(file)

plot(
  x = out,
  begin = 0,
  end = 10,
  panel_height = 1,
  zero_line = FALSE,
  small_squares = TRUE
)

}
