% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.empi.R
\name{plot.empi}
\alias{plot.empi}
\title{Plots a time-frequency (T-F) map to visualize EMPI decomposition}
\usage{
\method{plot}{empi}(
  x,
  channel = 1,
  mode = "sqrt",
  freq.divide = NULL,
  increase.factor = 8,
  shortening.factor.x = 2,
  shortening.factor.y = 2,
  display.crosses = TRUE,
  display.atom.numbers = FALSE,
  display.grid = FALSE,
  crosses.color = "white",
  palette = "my custom palette",
  plot.signals = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{empi} created by \code{empi.execute()}.}

\item{channel}{Channel from the SQLite file to process.}

\item{mode}{\code{"sqrt"}, \code{"log"}, or \code{"linear"}. Determines the intensity
with which the so-called blobs are displayed on the T-F map.}

\item{freq.divide}{Specifies how many times the displayed frequency range in the T-F map
should be reduced. At high sampling rates, and when a low-pass filter with
a cut-off frequency much lower than the sampling frequency is used, a large part of
the T-F map may contain no blobs. If the sampling frequency is \code{f},
the maximum frequency in the T-F map will be \code{ceiling(f / 2 / freq.divide)}
(\code{f / 2} follows the Nyquist rule). If \code{NULL}, it is determined from the atom
with the highest frequency \code{fmax} according to \code{freq.divide = (f / 2) / fmax}.}

\item{increase.factor}{Factor controlling the increase in the number of pixels along the
frequency axis. Non-negative integers such as 2, 4, 5, or 8 are typically appropriate.}

\item{shortening.factor.x}{Usually, a value of 2 provides better visualization of atoms.}

\item{shortening.factor.y}{Usually, a value of 2 provides better visualization of atoms.}

\item{display.crosses}{Whether small crosses should be displayed at the centres of atoms.}

\item{display.atom.numbers}{Whether atom numbers should be displayed at the centres of atoms.}

\item{display.grid}{Whether grid lines should be drawn.}

\item{crosses.color}{Colour of the small crosses.}

\item{palette}{Palette from the list returned by \code{hcl.pals()} or the string
\code{"my custom palette"}.}

\item{plot.signals}{Whether the original and reconstructed signals should also be displayed.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize the empi decomposition.
}
\description{
This function is a wrapper around \code{empi2tf()} with \code{out.mode = "plot"}.
}
\examples{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
signal <- read.csv.signals(file, col.names = "ch1")

# Execute the MP algorithm.
empi.class <- empi.execute(signal = signal)

# Plot a time-frequency map based on MP atoms.
plot(empi.class)

}
