% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mp.R
\name{plot.mp}
\alias{plot.mp}
\title{Plots a time-frequency (T-F) map to visualize EMPI decomposition}
\usage{
\method{plot}{mp}(
  x,
  channel = 1,
  mode = "sqrt",
  freq_divide = NULL,
  increase_factor = 8,
  shortening_factor_x = 2,
  shortening_factor_y = 2,
  display_crosses = TRUE,
  display_atom_numbers = FALSE,
  display_grid = FALSE,
  color = "white",
  palette = "my custom palette",
  plot_signals = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{mp} returned by \code{empi_execute()} or \code{mp_omp_execute()}.}

\item{channel}{Channel from the SQLite file to process.}

\item{mode}{\code{"sqrt"}, \code{"log"}, or \code{"linear"}. Determines the intensity
with which the so-called blobs are displayed on the T-F map.}

\item{freq_divide}{Specifies how many times the displayed frequency range in the T-F map
should be reduced. At high sampling rates, and when a low-pass filter with
a cut-off frequency much lower than the sampling frequency is used, a large part of
the T-F map may contain no blobs. If the sampling frequency is \code{f},
the maximum frequency in the T-F map will be \code{ceiling(f / 2 / freq_divide)}
(\code{f / 2} follows the Nyquist rule). If \code{NULL}, it is determined from the atom
with the highest frequency \code{fmax} according to \code{freq_divide = (f / 2) / fmax}.}

\item{increase_factor}{Factor controlling the increase in the number of pixels along the
frequency axis. Non-negative integers such as 2, 4, 5, or 8 are typically appropriate.}

\item{shortening_factor_x}{Usually, a value of 2 provides better visualization of atoms.}

\item{shortening_factor_y}{Usually, a value of 2 provides better visualization of atoms.}

\item{display_crosses}{Whether small crosses should be displayed at the centres of atoms.}

\item{display_atom_numbers}{Whether atom numbers should be displayed at the centres of atoms.}

\item{display_grid}{Whether grid lines should be drawn.}

\item{color}{Color of the small crosses and atom numbers}

\item{palette}{Palette from the list returned by \code{hcl.pals()} or the string
\code{"my custom palette"}.}

\item{plot_signals}{Whether the original and reconstructed signals should also be displayed.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize the EMPI decomposition.
}
\description{
This function is a wrapper around \code{tf_map()} with \code{out_mode = "plot"}.
}
\examples{
\dontrun{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
signal <- read_csv_signals(file, col_names = "ch1")

# Execute the MP algorithm.
out_empi <- empi_execute(signal = signal)

# Plot a time-frequency map based on MP atoms.
plot(out_empi)
}

}
