% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wfdb.R
\name{plot.wfdb}
\alias{plot.wfdb}
\title{The function displays WFDB signals in a layout corresponding to standard paper ECG printouts}
\usage{
\method{plot}{wfdb}(
  x,
  begin = NULL,
  end = NULL,
  panel_height = 3,
  small_squares = TRUE,
  zero_line = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{wfdb} (from \code{read_wfdb_signals()}).}

\item{begin}{Time point (in seconds) at which to start plotting.
If \code{NULL}, plotting starts at the beginning of the signal (0 seconds).}

\item{end}{Time point (in seconds) at which to stop plotting.
If \code{NULL}, plotting continues to the end of the signal.}

\item{panel_height}{Height of each ECG channel panel (in mV).
One large ECG-paper square corresponds to 0.5 mV. According to standard ECG paper:
\itemize{
    \item small grid: 0.04 sec. x 0.1 mV
    \item large grid: 0.20 sec. x 0.5 mV
}}

\item{small_squares}{If \code{TRUE}, the small grid is also displayed.}

\item{zero_line}{If \code{TRUE}, a horizontal line representing \code{0 mV} is displayed.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize an ECG graph.
}
\description{
ECG signals are read from files in WFDB format.
}
\details{
WFDB (WaveForm DataBase) is a standard file format for storing, reading,
and analyzing physiological time-series signals. It is widely used for
signals such as ECG, EEG, blood pressure, respiration, and other biomedical
waveforms. It is the file format used by the PhysioNet project and is commonly
used in research datasets.

A WFDB record typically consists of two main files:
\code{.dat} - binary signal samples (waveform values), and \code{.hea} - a header
file describing how to interpret the data. In some cases, additional annotation
files such as \code{.atr} may be present, containing beat labels or rhythm annotations.

A typical ECG paper layout was used, with a small grid of 0.04 s × 0.1 mV and a
large grid of 0.20 s × 0.5 mV.
}
\examples{
# ECG data comes from https://physionet.org/content/ptb-xl/1.0.3/
file <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")
out <- read_wfdb_signals(file)

plot(
  x = out,
  begin = 0,
  end = 10,
  panel_height = 1,
  zero_line = FALSE,
  small_squares = TRUE
)

}
