% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ecg.signals.R
\name{read.ecg.signals}
\alias{read.ecg.signals}
\title{Reads WFDB-compatible signal & header files}
\usage{
read.ecg.signals(file)
}
\arguments{
\item{file}{The path to the ECG record to be read.}
}
\value{
A list is returned with:
1) data frame with all signals stored in the given ECG file,
3) sampling rate,
4) lead names,
4) object of class \code{ecg}.
}
\description{
WFDB (WaveForm DataBase) is a standard file format for storing, reading,
and analyzing physiological time-series signals.It is widely used for signals
such as: ECG, EEG, blood pressure, respiration and other biomedical waveforms.
It was developed by PhysioNet and is common in research datasets.
A WFDB record usually contains two main files: \code{.dat} — binary signal
samples (the waveform values) and \code{.hea} — header file describing how to
interpret the data. Sometimes there are also annotation files such as
\code{.atr}, containing beat labels or rhythm annotations.
}
\examples{
# ECG data comes from https://physionet.org/content/ptb-xl/1.0.3/
file <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")
dir <- dirname(file)
name <- tools::file_path_sans_ext(basename(file))

out <- read.ecg.signals(file)
head(out$signals)
out$sampling.rate
out$lead.names

}
