% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.edf.params.R
\name{read.edf.params}
\alias{read.edf.params}
\title{Reads a selected EDF or EDF+ file and returns signal parameters}
\usage{
read.edf.params(file)
}
\arguments{
\item{file}{The path to the EDF / EDF+ file to be read.}
}
\value{
A data frame is returned containing the most basic parameters of the EDF / EDF+ file.
}
\description{
Reads a selected EDF or EDF+ file and returns selected signals parameters
(channel names, frequency of each channel, number of samples in each channel
and the length of each channel in seconds). Additional information stored in EDF+
files (such as interrupted recordings, time-stamped annotations) is not used in the
package and is therefore not read.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
read.edf.params(file)

}
