% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.empi.db.file.R
\name{read.empi.db.file}
\alias{read.empi.db.file}
\title{Reads data from a SQLite file created by the Matching Pursuit algorithm}
\usage{
read.empi.db.file(db.file)
}
\arguments{
\item{db.file}{SQLite file.}
}
\value{
\itemize{
   \item Detailed parameters of all the generated atoms
   \item Original input signal(s)
   \item Reconstructed signal(s), as the sum of generated atoms
   \item Generated Gabor atoms
   \item time stamps
   \item sampling rate
}
}
\description{
Reads data from a SQLite file (\code{.db}) created by the Matching Pursuit algorithm.
The reconstructed signal(s) and Gabor function(s) are also returned.
}
\examples{
## Not run:
file <- system.file("extdata", "EEG.db", package = "MatchingPursuit")
out <- read.empi.db.file(file)

n.channels <- ncol(out$original.signal)
original.signal <- out$original.signal
reconstruction <- out$reconstruction
t <- out$t
f <- out$f

old.par <- par("mfrow", "pty", "mai")

par(mfrow = c(2, 1))
par(pty = "m")
par(mai = c(0.9, 0.5, 0.3, 0.4))

plot(
  original.signal[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n.channels, " (original signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

len <- length(original.signal[, 1])
lab <- seq(t[1], t[len] + 1 / f, length.out = 11)
axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)

plot(
  reconstruction[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n.channels, " (reconstructed signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)

par(old.par)

## End(Not run)

}
