% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_atom_params.R
\name{read_atom_params}
\alias{read_atom_params}
\title{Read atom parameters from a SQLite database}
\usage{
read_atom_params(db_file)
}
\arguments{
\item{db_file}{A character string giving the path to a SQLite database file.}
}
\value{
A data frame containing the atom parameters stored in the database:

\item{channel_id}{Channel identifier.}
\item{atom_number}{Atom number.}
\item{energy}{Energy of the atom.}
\item{frequency}{Frequency of the atom.}
\item{phase}{Phase of the atom.}
\item{scale}{Scaling factor.}
\item{position}{Position of the atom in time.}
}
\description{
Reads the atom parameters from a SQLite database produced by \code{empi_execute()}.
}
\examples{
# Example database containing data from 18 channels
file <- system.file("extdata", "EEG_filter_resample_montage.db", package = "MatchingPursuit")
out <- read_atom_params(file)
out[which(out$channel_id == 1), ]
out[which(out$channel_id == 18), ]

}
