% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecg_signals.R
\name{read_ecg_signals}
\alias{read_ecg_signals}
\title{Reads WFDB-compatible signal and header files}
\usage{
read_ecg_signals(file)
}
\arguments{
\item{file}{Path to the ECG record to be read.}
}
\value{
An object of class \code{ecg}. The returned value is a list containing:

\item{signal}{Matrix of signals stored in the ECG file.}
\item{sampling_frequency}{Sampling frequency.}
\item{time_stamps}{Time vector corresponding to signal samples.}
\item{lead_names}{Names of the ECG leads (channels).}
\item{record_name}{Name of the file.}
}
\description{
WFDB (WaveForm DataBase) is a standard file format for storing, reading,
and analyzing physiological time-series signals. It is widely used for
signals such as ECG, EEG, blood pressure, respiration, and other biomedical
waveforms. It was developed by PhysioNet and is commonly used in research datasets.
}
\details{
A WFDB record typically consists of two main files:
\code{.dat} - binary signal samples (waveform values), and \code{.hea} - a header
file describing how to interpret the data. In some cases, additional annotation
files such as \code{.atr} may be present, containing beat labels or rhythm annotations.
}
\examples{
# ECG data comes from https://physionet.org/content/ptb-xl/1.0.3/
file <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")
dir <- dirname(file)
name <- tools::file_path_sans_ext(basename(file))

out <- read_ecg_signals(file)
head(out$signal)
out$sampling_frequency
out$lead_names

plot(out, begin = 0, end = 10, panel_height = 1.5)
}
