% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edf_params.R
\name{read_edf_params}
\alias{read_edf_params}
\title{Reads a selected EDF or EDF+ file and returns signal parameters}
\usage{
read_edf_params(file)
}
\arguments{
\item{file}{Path to the EDF / EDF+ file to be read.}
}
\value{
A data frame containing the basic parameters of the EDF / EDF+ file:

 \item{channel_name}{Name of the given channel.}
 \item{frequency}{Sampling frequency of the given channel.}
 \item{no_of_samples}{Number of samples in the given channel.}
 \item{length_sec}{Length in seconds of the given channel.}
}
\description{
Reads a selected EDF or EDF+ file and returns basic signal parameters
(channel names, sampling frequency of each channel, number of samples per channel,
and signal duration in seconds). Additional information stored in EDF+ files
(such as interrupted recordings or time-stamped annotations) is not used by the
package and is therefore not read.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
read_edf_params(file)

}
