% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_empi_db.R
\name{read_empi_db}
\alias{read_empi_db}
\title{Read EMPI decomposition results from a SQLite database}
\usage{
read_empi_db(db_file)
}
\arguments{
\item{db_file}{A character string giving the path to a SQLite database file.}
}
\value{
An object of class \code{"mp"} containing:
\describe{
  \item{atoms}{A data frame describing the selected atoms.}
  \item{signal}{Matrix containing the original signal(s).}
  \item{reconstruction}{Matrix containing the reconstructed signal(s).}
  \item{selected_atoms}{List of matrices containing selected atoms for each channel.}
  \item{time}{Time vector corresponding to signal samples.}
  \item{sampling_frequency}{Sampling frequency.}
}
}
\description{
Reads data from a SQLite file (\code{.db}) created by the Matching Pursuit algorithm.
The reconstructed signal(s) and Gabor function(s) are also returned.
}
\examples{
file <- system.file("extdata", "EEG_filter_resample_montage.db", package = "MatchingPursuit")
out <- read_empi_db(file)

n_channels <- ncol(out$signal)
signal <- out$signal
reconstruction <- out$reconstruction
t <- out$time
sampling_frequency <- out$sampling_frequency

old.par <- par("mfrow", "pty", "mai")

par(mfrow = c(2, 1))
par(pty = "m")
par(mai = c(0.9, 0.5, 0.3, 0.4))

plot(
  signal[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n_channels, " (original signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

len <- length(signal[, 1])
lab <- seq(t[1], t[len] + 1 / sampling_frequency, length.out = 11)
axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)

plot(
  reconstruction[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n_channels, " (reconstructed signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)

par(old.par)

}
