% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wfdb_signals.R
\name{read_wfdb_signals}
\alias{read_wfdb_signals}
\title{Reads WFDB-compatible signal and header files}
\usage{
read_wfdb_signals(file)
}
\arguments{
\item{file}{Path to the WFDB record to be read.}
}
\value{
An object of class \code{wfdb}. The returned value is a list containing:

\describe{
  \item{signal}{Matrix of signals stored in the WFDB file.}
  \item{sampling_frequency}{Sampling frequency.}
  \item{time}{Time vector corresponding to signal samples.}
  \item{lead_names}{Names of the WFDB leads (channels).}
  \item{record_name}{Name of the file.}
}
}
\description{
WFDB (WaveForm DataBase) is a standard file format for storing, reading,
and analyzing physiological time-series signals. It is widely used for
signals such as ECG, EEG, blood pressure, respiration, and other biomedical
waveforms. It is the file format used by the PhysioNet project and is commonly
used in research datasets.
}
\details{
A WFDB record typically consists of two main files:
\code{.dat} - binary signal samples (waveform values), and \code{.hea} - a header
file describing how to interpret the data. In some cases, additional annotation
files such as \code{.atr} may be present, containing beat labels or rhythm annotations.
}
\note{
The function \code{EGM::read_wfdb()} from version 0.2.0 of the
\code{EGM} package does not support multi-frequency signals. Consequently,
records containing different numbers of samples per frame, as indicated by
the \code{16x2}, \code{16x4}, and \code{16x1} specifications below, cannot
be read correctly.

\verb{
multi_freq_test 3 100 1000
multi_freq_test.dat 16x2 200.0(0)/mV 16 0 0 258 0 ECG
multi_freq_test.dat 16x4 400.0(0)/mmHg 16 0 400 57824 0 ABP
multi_freq_test.dat 16x1 100.0(0)/pm 16 0 0 18204 0 RESP
}
}
\examples{
# ECG data comes from https://physionet.org/content/ptb-xl/1.0.3/
file <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")

out <- read_wfdb_signals(file)
head(out$signal)
out$sampling_frequency
out$lead_names

plot(out, begin = 0, end = 10, panel_height = 1.5)

}
