% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_omp_pipeline.R
\name{run_omp_pipeline}
\alias{run_omp_pipeline}
\title{Run the complete OMP decomposition pipeline}
\usage{
run_omp_pipeline(
  sig_file,
  col_names_in_csv,
  xml_file,
  topk,
  n_nonzero_coefs,
  tol = NULL,
  normalize = TRUE,
  fit_intercept = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sig_file}{Path to a CSV file containing the signal data.
See \code{read_csv_signals()} function.}

\item{col_names_in_csv}{Logical. Indicates whether the CSV file contains column names in the first row.
See \code{read_csv_signals()} function.}

\item{xml_file}{Path to an XML file defining the dictionary atoms.
See \code{read_dict()} function.}

\item{topk}{Integer. Number of atoms with the highest correlation to the
signal retained for OMP fitting.
See \code{topk_atoms()} function.}

\item{n_nonzero_coefs}{Integer. Maximum number of non-zero coefficients  to include in the sparse decomposition.}

\item{tol}{Optional numeric tolerance for the stopping criterion. If \code{NULL}, the
algorithm stops after selecting \code{n_nonzero_coefs} atoms.}

\item{normalize}{Logical. If \code{TRUE}, dictionary atoms are normalized  before fitting.}

\item{fit_intercept}{Logical. If  \code{TRUE}, an intercept term is included in the model.}

\item{verbose}{Logical. If  \code{TRUE}, progress information is printed to  the console.}
}
\value{
An object of class \code{"mp"} containing:

\item{atoms}{A data frame describing the selected atoms.}
\item{original_signal}{Matrix containing the original signal(s).}
\item{reconstruction}{Matrix containing the reconstructed signal(s).}
\item{gabors}{List of matrices containing selected atoms for each channel.}
\item{t}{Time vector corresponding to signal samples.}
\item{sf}{Sampling frequency.}
}
\description{
Executes the full Orthogonal Matching Pursuit (OMP) workflow:
reads a signal from a CSV file, loads a dictionary definition from an
XML file, selects the most relevant atoms, and performs sparse signal
decomposition using OMP.
}
\examples{
sig_file <- system.file("extdata", "sample3.csv", package = "MatchingPursuit")
xml_file <- system.file("extdata", "sample3_dict.xml", package = "MatchingPursuit")

# set 'verbose = TRUE' to see the progress

out <- run_omp_pipeline(
  sig_file = sig_file,
  col_names_in_csv = TRUE,
  xml_file = xml_file,
  topk = 5000,
  n_nonzero_coefs = 50,
  verbose = TRUE
)

plot(out, channel = 3)

}
\seealso{
\code{\link{omp_core}},
\code{\link{read_dict}},
\code{\link{omp_execute}}
\code{\link{topk_atoms}}
}
