% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-methods.R
\name{sig-methods}
\alias{sig-methods}
\alias{print.sig}
\alias{summary.sig}
\alias{print.summary.sig}
\alias{plot.sig}
\title{Methods for Signal Objects}
\usage{
\method{print}{sig}(x, ...)

\method{summary}{sig}(object, ...)

\method{print}{summary.sig}(x, ...)

\method{plot}{sig}(
  x,
  begin = NULL,
  end = NULL,
  panel_height = NULL,
  zero_line = TRUE,
  main = NULL,
  mar = c(4, 7, 2, 1),
  col = "black",
  lwd = 1,
  lty = 1,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"sig"} or \code{"summary.sig"}.}

\item{...}{Additional arguments. For \code{plot.sig()}, these are passed
to \code{plot()}; otherwise they are currently ignored.}

\item{object}{An object of class \code{"sig"}.}

\item{begin}{Beginning of the displayed interval in seconds.}

\item{end}{End of the displayed interval in seconds.}

\item{panel_height}{Vertical distance between signal channels. If \code{NULL},
it is determined automatically from the signal amplitude.}

\item{zero_line}{Logical. If \code{TRUE}, draw a horizontal reference line
for each channel.}

\item{main}{Optional plot title.}

\item{mar}{Numeric vector of length four specifying the plot margins in the
form \code{c(bottom, left, top, right)}.}

\item{col}{Colour used to draw the signal traces.}

\item{lwd}{Line width used to draw the signal traces.}

\item{lty}{Line type used to draw the signal traces.}
}
\value{
\code{print.sig()} and \code{print.summary.sig()} return their input
object invisibly.

\code{summary.sig()} returns an object of class \code{"summary.sig"}
containing signal dimensions, sampling information, channel names,
and basic descriptive statistics for each channel.

\code{plot.sig()} is called for its side effect and returns \code{x} invisibly.
}
\description{
Methods for printing, summarizing, and plotting objects of class
\code{"sig"}, together with a print method for objects returned by
\code{summary()}.
}
\details{
Objects of class \code{"sig"} represent a single- or multi-channel signal
together with its sampling frequency and corresponding time vector.
}
\examples{
file <- system.file("extdata", "sample3.csv", package = "MatchingPursuit")

x <- read_csv_signals(
  file,
  col_names_in_csv = TRUE
)

print(x)
summary(x)

plot(x)
plot(x, mar = c(4, 12, 2, 1))
plot(x, mar = c(4, 12, 2, 1), col = "blue")
plot(x, col = "red", lwd = 2, lty = 4)
plot(x, begin = 0, end = 1, main = "Signal")

}
\seealso{
\code{\link{read_csv_signals}},
\code{\link{as_sig}}
}
