% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom.params.R
\name{atom.params}
\alias{atom.params}
\title{Read atom parameters from a SQLite database}
\usage{
atom.params(db.file)
}
\arguments{
\item{db.file}{A character string giving the path to a SQLite database file.}
}
\value{
A data frame containing the atom parameters stored in the database.
}
\description{
Reads atom parameters stored in a SQLite database created by \code{empi.execute()} function.
}
\examples{
# Example database containing data from 18 channels
file <- system.file("extdata", "EEG.db", package = "MatchingPursuit")
out <- atom.params(file)
out[which(out$channel_id == 1), ]
out[which(out$channel_id == 18), ]

# Example database containing data from a single channel
file <- system.file("extdata", "sample1.db", package = "MatchingPursuit")
out <- atom.params(file)
out

}
