% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom.params.R
\name{atom.params}
\alias{atom.params}
\title{Returns data frame with all the atoms' parameters.}
\usage{
atom.params(db.file)
}
\arguments{
\item{db.file}{The SQLite file created after executing the \code{empi.execute} function.}
}
\value{
Data frame with all the atoms' parameters saved in a given SQLite file.
}
\description{
Returns data frame with all the atoms' parameters.
}
\examples{
file <- system.file("extdata", "EEG_data_10sec.db", package = "MatchingPursuit")
out <- atom.params(file)
out[which(out$channel_id == 1), ]
out[which(out$channel_id == 18), ]

}
