% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom.params.R
\name{atom.params}
\alias{atom.params}
\title{Reading the atom parameters}
\usage{
atom.params(db.file)
}
\arguments{
\item{db.file}{The SQLite file created after executing the \code{empi.execute} function.}
}
\value{
Data frame with all the atoms' parameters saved in a given SQLite file. The file can be generated using the \code{empi.execute} function.
}
\description{
Returns a data frame with atom parameters.
}
\examples{
# The file contains data from 18 channels.
file <- system.file("extdata", "EEG_data_10sec.db", package = "MatchingPursuit")
out <- atom.params(file)
out[which(out$channel_id == 1), ]
out[which(out$channel_id == 18), ]

# Only 1 channel is in the file.
file <- system.file("extdata", "sample1.db", package = "MatchingPursuit")
out <- atom.params(file)
out

}
