% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom_params.R
\name{atom_params}
\alias{atom_params}
\title{Read atom parameters from a SQLite database}
\usage{
atom_params(db_file)
}
\arguments{
\item{db_file}{A character string giving the path to a SQLite database file.}
}
\value{
A data frame containing the atom parameters stored in the database:

\item{channel_id}{Channel identifier.}
\item{atom_number}{Atom number.}
\item{energy}{Energy of the atom.}
\item{frequency}{Frequency of the atom.}
\item{phase}{Phase of the atom.}
\item{scale}{Scaling factor.}
\item{position}{Position of the atom in time.}
}
\description{
Reads atom parameters stored in a SQLite database created by \code{empi_execute()} function.
}
\examples{
# Example database containing data from 18 channels
file <- system.file("extdata", "EEG_bipolar_filtered.db", package = "MatchingPursuit")
out <- atom_params(file)
out[which(out$channel_id == 1), ]
out[which(out$channel_id == 18), ]

# Example database containing data from a single channel
file <- system.file("extdata", "sample1.db", package = "MatchingPursuit")
out <- atom_params(file)
out

}
