% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.check.R
\name{empi.check}
\alias{empi.check}
\title{Checks if EMPI external software is installed}
\usage{
empi.check()
}
\value{
If the EMPI program is found, its full path is returned. Otherwise, a message is displayed,
prompting the user to install it using the \code{empi.install()} function.
}
\description{
The EMPI program is installed using the \code{empi.install()} function and is stored in the
cache directory. This function checks whether the EMPI program is still there (the user has
free access to the cache directory and can, for example, delete it at any time).
}
\examples{
empi.check()

}
