% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.execute.R
\name{empi.execute}
\alias{empi.execute}
\title{Runs the \emph{empi} program for the given data.}
\usage{
empi.execute(signal, sampling.rate, empi.options = NULL, write.to.file = TRUE)
}
\arguments{
\item{signal}{Must be a data frame: rows = samples for all channels, columns = channels. The data frame must have column names (channel names).}

\item{sampling.rate}{Sampling rate of the given signal (must be the same for every channel).}

\item{empi.options}{If NULL, the \emph{empi} program runs with \code{"-o local --gabor -i 50"}. Otherwise, user can spacify any command-line options. See \code{README.md} file after downloading the \emph{empi} program using \code{empi.download} function.}

\item{write.to.file}{If TRUE, the SQLite database file \code{empi.db} will be created and saved in the current directory. This file stores the results of signal decomposition using the Matching Pursuit algorithm.}
}
\value{
Results of signal decomposition using the Matching Pursuit algorithm. If \code{write.to.file=TRUE}, additionally, these results are written to the \code{empi.db} disk file in the working directory.
}
\description{
todo.
}
\examples{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
signal <- read.csv(file, header = TRUE)

empi.out <- empi.execute (
  signal = signal,
  sampling.rate = 128,
  empi.options = NULL,
  write.to.file = FALSE
)

}
