% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.execute.R
\name{empi.execute}
\alias{empi.execute}
\title{Launching the Matching Pursuit algorithm}
\usage{
empi.execute(signal, sampling.rate, empi.options = NULL, write.to.file = FALSE)
}
\arguments{
\item{signal}{Must be a data frame: rows = samples for all channels, columns = channels.
The data frame must have column names (channel names).}

\item{sampling.rate}{Sampling rate of the given signal (must be the same for every channel).}

\item{empi.options}{If NULL, the \emph{empi} program runs with \code{"-o local --gabor -i 50"}.
Otherwise, user can spacify any command-line options. See \code{README.md} file after downloading
the \emph{empi} program using \code{empi.download()} function.}

\item{write.to.file}{If TRUE, the SQLite database file \code{empi.db} will be created and saved
in the current directory. This file stores the results of signal decomposition using the
Matching Pursuit algorithm.}
}
\value{
Results of signal decomposition using the Matching Pursuit algorithm. If
\code{write.to.file=TRUE} additionally, these results are written to the \code{empi.db} disk
file in the working directory.
}
\description{
Runs the \emph{empi} program for the given data.
}
\details{
The \emph{empi} program (sources and binaries for different OSs) is available for download at
\url{https://github.com/develancer/empi}. Details are presented in a journal paper:
Różański, P.T. (2024). \emph{empi: GPU-Accelerated Matching Pursuit with Continuous Dictionaries}.
ACM Transactions on Mathematical Software, vol.50, issue = 3, pp. 1-17,
\url{https://doi.org/10.1145/3674832}.
}
\examples{
## Not run:
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
signal <- read.csv(file, header = TRUE)

if (interactive()) {
empi.out <- empi.execute (
  signal = signal,
  sampling.rate = 128,
  empi.options = NULL,
  write.to.file = FALSE
)
}
## End(Not run)

}
