% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.install.R
\name{empi.install}
\alias{empi.install}
\title{Installs the required external program}
\usage{
empi.install()
}
\value{
The function downloads the EMPI program in a version compatible with the operating
system used (Windows, Linux, MacOS-x64, MacOS-arm64) and stores it in the cache directory.
}
\description{
Downloads \strong{Enhanced Matching Pursuit Implementation} external program (or EMPI for short)
and stores it in the cache directory.
}
\examples{
if (interactive()) {
  empi.install()
}

}
