% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi.locate.R
\name{empi.locate}
\alias{empi.locate}
\title{Get required external software localization}
\usage{
empi.locate()
}
\value{
List with URL of the \emph{empi} binaries and zip file name.
}
\description{
Returns \emph{Enhanced Matching Pursuit Implementation} \url{https://github.com/develancer/empi}
binary locations for the following operation systems: Windows, Linux, MacOS-x64, MacOS-arm64.
}
\examples{
empi.locate()

}
