% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi2tf.R
\name{empi2tf}
\alias{empi2tf}
\title{Creates a time-frequency map using atoms from the Matching Pursuit algorithm}
\usage{
empi2tf(
  db.file = NULL,
  db.list = NULL,
  channel,
  mode = "sqrt",
  freqDivide = 1,
  increaseFactor = 1,
  shorteningFactor.x = 2,
  shorteningFactor.y = 2,
  displayCrosses = TRUE,
  displayAtomNumbers = FALSE,
  displayGrid = FALSE,
  crossesColor = "white",
  palette = "my custom palette",
  rev = TRUE,
  outMode = "file",
  fileName = "sample.png",
  fileSize = c(512, 512),
  drawEllipses = FALSE,
  plotSignals = TRUE,
  plotAtoms = FALSE
)
}
\arguments{
\item{db.file}{The SQLite file created after executing the \code{empi.execute} function. In this case, the \code{db.list} parameter must be NULL.}

\item{db.list}{The list created after executing the \code{empi.execute} function. In this case, the \code{db.file} parameter must be NULL.}

\item{channel}{Channel from the SQLite file to process.}

\item{mode}{\code{"sqrt"}, \code{"log"}, or \code{"linear"}.}

\item{freqDivide}{For setting frequency range. For example, when sampling frequency is \code{f=256Hz}, the maximum frequency is \code{f/2} (Nyquist rule) and the limited the frequency is  \code{f/2/freqDivide}.}

\item{increaseFactor}{Factor of increasing the number of pixels in the f-axis, the most sensible are non-negative integers (e.g. 2, 4, 5, 8).}

\item{shorteningFactor.x}{Usually, for better visualization of atoms, a value of 2 will be appropriate.}

\item{shorteningFactor.y}{Usually, for better visualization of atoms, a value of 2 will be appropriate.}

\item{displayCrosses}{Whether small crosses should be displayed in the canters of atoms.}

\item{displayAtomNumbers}{Whether atom numbers should be displayed in the canters of atoms.}

\item{displayGrid}{Whether to draw grid lines.}

\item{crossesColor}{Colour of small crosses.}

\item{palette}{Palette from the list returned by \code{hcl.pals} function or the string \code{"my custom palette"}.}

\item{rev}{\code{rev} param in \code{hcl.colors} function.}

\item{outMode}{One of the following:
\itemize{
   \item \code{"plot"} - draws a TF map on the screen.
   \item \code{"file"} - saves a TF map to file \code{fileName} (as png file).
   \item \code{"RData"} - saves the TF map of \code{fileSize} in the \code{fileName} (as R's matrix) resampling using \code{imager::resize} function.
   \item \code{"RData2"} - saves the TF map of \code{fileSize} in the \code{fileName} (as R's matrix) resampling using \code{raster::resample} function.
 }}

\item{fileName}{Name of the png file.}

\item{fileSize}{File size in pixels.}

\item{drawEllipses}{Only for testing.}

\item{plotSignals}{Whether the original and reconstructed signals should also be displayed.}

\item{plotAtoms}{If TRUE, plot all atoms into \code{Atoms.pdf} file.}
}
\value{
Depending on the \code{outMode} parameter the function returns:
   \itemize{
   \item a time-Frequency map plotted on the screen.
   \item a time-Frequency map saved in a png file.
   \item a time-Frequency map saved as .RData file.
  }
}
\description{
The map can be displayed on the screen, saved to a \code{png} file, or saved as an RData object (as a matrix).
}
\examples{
file <- system.file("extdata", "sample2.db", package = "MatchingPursuit")

out <- empi2tf(
  db.file = file,
  channel = 1,
  mode = "sqrt",
  freqDivide = 1,
  increaseFactor= 4,
  shorteningFactor.x = 2,
  shorteningFactor.y = 2,
  displayCrosses = TRUE,
  displayAtomNumbers = FALSE,
  outMode = "plot"
)

}
