% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi2tf.R
\name{empi2tf}
\alias{empi2tf}
\title{Creates a time-frequency map using atoms from the Matching Pursuit algorithm}
\usage{
empi2tf(
  db.file = NULL,
  db.list = NULL,
  channel,
  mode = "sqrt",
  freq.divide = 1,
  increase.factor = 1,
  shortening.factor.x = 2,
  shortening.factor.y = 2,
  display.crosses = TRUE,
  display.atom.numbers = FALSE,
  display.grid = FALSE,
  crosses.color = "white",
  palette = "my custom palette",
  rev = TRUE,
  out.mode = "file",
  file.name = "sample.png",
  file.size = c(512, 512),
  draw.ellipses = FALSE,
  plot.signals = TRUE,
  plot.atoms = FALSE
)
}
\arguments{
\item{db.file}{The SQLite file created after executing the \code{empi.execute()} function.
In this case, the \code{db.list} parameter must be NULL.}

\item{db.list}{The list created after executing the \code{empi.execute()} function.
In this case, the \code{db.file} parameter must be NULL.}

\item{channel}{Channel from the SQLite file to process.}

\item{mode}{\code{"sqrt"}, \code{"log"}, or \code{"linear"}.}

\item{freq.divide}{For setting frequency range. For example, when sampling frequency is
\code{f=256Hz}, the maximum frequency is \code{f/2} (Nyquist rule) and the limited the
frequency is  \code{f/2/freq.divide}.}

\item{increase.factor}{Factor of increasing the number of pixels in the f-axis, the most
sensible are non-negative integers (e.g. 2, 4, 5, 8).}

\item{shortening.factor.x}{Usually, for better visualization of atoms, a value of 2 will be appropriate.}

\item{shortening.factor.y}{Usually, for better visualization of atoms, a value of 2 will be appropriate.}

\item{display.crosses}{Whether small crosses should be displayed in the canters of atoms.}

\item{display.atom.numbers}{Whether atom numbers should be displayed in the canters of atoms.}

\item{display.grid}{Whether to draw grid lines.}

\item{crosses.color}{Colour of small crosses.}

\item{palette}{Palette from the list returned by \code{hcl.pals()} function or the string
\code{"my custom palette"}.}

\item{rev}{\code{rev} param in \code{hcl.colors()} function.}

\item{out.mode}{One of the following:
\itemize{
   \item \code{"plot"} - draws a TF map on the screen.
   \item \code{"file"} - saves a TF map to file \code{file.name} (as png file).
   \item \code{"RData"} - saves the TF map of \code{file.size} in the \code{file.name}
   (as R's matrix) resampling using \code{imager::resize()} function.
   \item \code{"RData2"} - saves the TF map of \code{file.size} in the \code{file.name}
   (as R's matrix) resampling using \code{raster::resample()} function.
 }}

\item{file.name}{Name of the png file.}

\item{file.size}{File size in pixels.}

\item{draw.ellipses}{Only for testing.}

\item{plot.signals}{Whether the original and reconstructed signals should also be displayed.}

\item{plot.atoms}{If TRUE, plot all atoms into \code{Atoms.pdf} file.}
}
\value{
Depending on the \code{out.mode} parameter the function returns:
   \itemize{
   \item a time-Frequency map plotted on the screen
   \item a time-Frequency map saved in a png file
   \item a time-Frequency map saved as .RData file
  }
Regardless of the above, the function returns the following:
  \itemize{
    \item all the Gabor function
    \item reconstructed signal
    \item original signal
    \item sampling frequency
    \item grid size in t axis
    \item grid size in f axi
    \item epch size in samples
    \item length of the signal in seconds
    \item fime-frequency map
    \item channel number
    \item frequency divide
  }
}
\description{
The map can be displayed on the screen, saved to a \code{png} file, or saved as an RData object
(as a matrix).
}
\examples{
file <- system.file("extdata", "sample2.db", package = "MatchingPursuit")

out <- empi2tf(
  db.file = file,
  channel = 1,
  mode = "sqrt",
  freq.divide = 1,
  increase.factor= 4,
  shortening.factor.x = 2,
  shortening.factor.y = 2,
  display.crosses = TRUE,
  display.atom.numbers = FALSE,
  out.mode = "plot"
)

}
