% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi_check.R
\name{empi_check}
\alias{empi_check}
\title{Checks if EMPI external software is installed}
\usage{
empi_check()
}
\value{
The full path to the EMPI executable if it is found. Otherwise, returns \code{NULL}
and displays a message suggesting installation using \code{empi_install()}.
}
\description{
The EMPI program is installed using the \code{empi_install()} function and stored in the
cache directory. This function checks whether the EMPI program is still available there
(users have full access to the cache directory and may remove its contents at any time).
}
\examples{
if (interactive()) {
  empi_check()
}

}
\seealso{
\code{\link{empi_install}},
\code{\link{empi_locate}},
\code{\link{empi_execute}},
\code{\link{plot.mp}}
}
