% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi_check.R
\name{empi_check}
\alias{empi_check}
\title{Checks if EMPI external software is installed}
\usage{
empi_check()
}
\value{
If the EMPI program is found, its full path is returned. Otherwise, a message is displayed,
prompting the user to install it using the \code{empi_install()} function.
}
\description{
The EMPI program is installed using the \code{empi_install()} function and stored in the
cache directory. This function checks whether the EMPI program is still available there
(users have full access to the cache directory and may remove its contents at any time).
}
\examples{
empi_check()

}
