% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi_execute.R
\name{empi_execute}
\alias{empi_execute}
\title{Launches the empi program}
\usage{
empi_execute(
  signal,
  empi_options = NULL,
  write_to_file = FALSE,
  path = NULL,
  file_name = NULL,
  ...
)
}
\arguments{
\item{signal}{An object of class \code{sig} returned by \code{read_csv_signals()},
an object of class \code{edf} returned by \code{read_edf_signals()},
or an object of class \code{wfdb} returned by \code{read_wfdb_signals()}.}

\item{empi_options}{If \code{NULL}, the EMPI program is run with
\code{"-o local --gabor -i 50"} parameters. Otherwise, the user may specify any command-line
options. See the \code{README.md} file after downloading the EMPI program using the
\code{empi_install()} function.}

\item{write_to_file}{If \code{TRUE}, a SQLite database file will be created
and saved in the \code{path} directory or, if \code{path = NULL}, in the
cache directory. This file stores the results of signal decomposition using the
EMPI program.}

\item{path}{Directory in which the SQLite database file will be saved.
If \code{NULL}, the file will be saved in the cache directory.}

\item{file_name}{Name of the file to create if \code{write_to_file = TRUE}.}

\item{...}{Additional arguments passed to \code{system()} when executing
EMPI, such as \code{ignore.stdout = TRUE} or \code{ignore.stderr = TRUE}.}
}
\value{
Results of signal decomposition using the MP algorithm. An object of class
\code{mp} is returned. If \code{write_to_file = TRUE}, the results are also written
to a SQLite file in the \code{path} directory.

\item{atoms}{A data frame describing the selected atoms.}
\item{signal}{Matrix containing the original signal(s).}
\item{reconstruction}{Matrix containing the reconstructed signal(s).}
\item{selected_atoms}{List of matrices containing selected atoms for each channel.}
\item{time}{Time vector corresponding to signal samples.}
\item{sampling_frequency}{Sampling frequency.}
}
\description{
Runs the EMPI program for the given data (signal).
}
\details{
The EMPI program (source code and binary files for multiple operating systems) can be
downloaded from \url{https://github.com/develancer/empi}. Details are presented in the
journal paper: Różański, P. T. (2024). \emph{empi: GPU-Accelerated Matching Pursuit with
Continuous Dictionaries}. ACM Transactions on Mathematical Software, Volume 50, Issue 3,
Article No. 17, pp. 1-17, \doi{10.1145/3674832}.
}
\examples{
\dontrun{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
signal <- read_csv_signals(file)

out_empi <- empi_execute(
  signal = signal,
  empi_options = NULL,
  write_to_file = FALSE,
  path = NULL,
  file_name = NULL
)

# Default EMPI options have been changed. For details, see the EMPI README.md file.
out_empi <- empi_execute(
  signal = signal,
  empi_options = "-o none --full-atoms-in-signal -i 50 --gabor",
  write_to_file = FALSE,
  path = NULL,
  file_name = NULL
)

plot(out_empi, freq_divide = 4)
}

}
\seealso{
\code{\link{empi_check}},
\code{\link{empi_install}},
\code{\link{empi_locate}},
\code{\link{plot.mp}}
}
