% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi_execute.R
\name{empi_execute}
\alias{empi_execute}
\title{Launches the empi program}
\usage{
empi_execute(
  signal,
  empi_options = NULL,
  write_to_file = FALSE,
  path = NULL,
  file_name = NULL
)
}
\arguments{
\item{signal}{List containing the signal in a data frame together with its sampling frequency.
The data frame should have meaningful column names (channel names).
The list must contain elements named \code{"signal"} and \code{"sampling_frequency"}.}

\item{empi_options}{If \code{NULL}, the EMPI program is run with
\code{"-o local --gabor -i 50 --cpu-workers 8"} parameters. Otherwise, the user may specify any command-line
options. See the \code{README.md} file after downloading the EMPI program using the
\code{empi_install()} function.}

\item{write_to_file}{If \code{TRUE}, a SQLite database file will be created
and saved in the \code{path} directory or, if \code{path = NULL}, in the
cache directory. This file stores the results of signal decomposition using the MP algorithm}

\item{path}{Directory in which the SQLite database file will be saved.
If \code{NULL}, the file will be saved in the cache directory.}

\item{file_name}{Name of the file to create if \code{write_to_file = TRUE}.}
}
\value{
Results of signal decomposition using the MP algorithm. An object of class
\code{mp} is returned. If \code{write_to_file = TRUE}, the results are also written
to a SQLite file in the \code{path} directory.

\item{atoms}{A data frame describing the selected atoms.}
\item{original_signal}{Matrix containing the original signal(s).}
\item{reconstruction}{Matrix containing the reconstructed signal(s).}
\item{gabors}{List of matrices containing selected atoms for each channel.}
\item{t}{Time vector corresponding to signal samples.}
\item{sf}{Sampling frequency.}
}
\description{
Runs the EMPI program for the given data (signal).
}
\details{
The EMPI program (source code and binary files for multiple operating systems) can be
downloaded from \url{https://github.com/develancer/empi}. Details are presented in the
journal paper: Różański, P. T. (2024). \emph{empi: GPU-Accelerated Matching Pursuit with
Continuous Dictionaries}. ACM Transactions on Mathematical Software, Volume 50, Issue 3,
Article No. 17, pp. 1-17, \doi{10.1145/3674832}.
}
\examples{
\dontrun{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
out <- read_csv_signals(file)

out_empi <- empi_execute(
  signal = out,
  empi_options = NULL,
  write_to_file = FALSE,
  path = NULL,
  file_name = NULL
)

plot(out_empi)
}

}
