% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empi_locate.R
\name{empi_locate}
\alias{empi_locate}
\title{Get required external software localization}
\usage{
empi_locate()
}
\value{
List with URL of the EMPI binaries and zip file name.
}
\description{
Returns \strong{Enhanced Matching Pursuit Implementation} binary locations for
the following operating systems: Windows, Linux, macOS-x64, macOS-arm64.
}
\examples{
empi_locate()

}
