% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabor.fun.R
\name{gabor.fun}
\alias{gabor.fun}
\title{Gabor function implementation}
\usage{
gabor.fun(
  number.of.samples,
  sampling.frequency,
  mean,
  phase,
  sigma,
  frequency,
  normalization = TRUE
)
}
\arguments{
\item{number.of.samples}{How many samples should the generated atom consist of?}

\item{sampling.frequency}{Sampling frequency.}

\item{mean}{Time position.}

\item{phase}{Phase.}

\item{sigma}{Scale / width of the Gaussian window.}

\item{frequency}{Frequency of the sinusoid.}

\item{normalization}{If \code{TRUE}, norm of the generated atom equals 1.}
}
\value{
List of 4 vectors with cosine, gauss, gabor and time waveforms of size \code{number.of.samples}.
}
\description{
Gabor function is a sinusoidal wave localized by a Gaussian envelope. In signal processing it is
widely used as a basic building block for representing signals that are localized in both time and
frequency. Matching Pursuit algorithm uses a redundant dictionary of the so called \emph{Gabor atoms}.
Gabor atoms are ideal for Matching Pursuit because they: 1) provide optimal time–frequency
localization, 2) represent oscillatory signals well, 3) enable adaptive time-frequency decomposition.
}
\examples{
number.of.samples <- 512
sampling.frequency <- 256.0
mean <- 1
phase <- pi
sigma <- 0.5
frequency <- 5.0
normalization = TRUE

out <- gabor.fun(
  number.of.samples,
  sampling.frequency,
  mean,
  phase,
  sigma,
  frequency,
  normalization
)

# If normalization = TRUE, norm of atom = 1, we can check it
crossprod(out$gabor)

plot(out$t, out$gabor, type = "l", xlab = "t", ylab = "gabor", panel.first = grid())

}
