% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabor_projection_fft.R
\name{gabor_projection_fft}
\alias{gabor_projection_fft}
\title{FFT-based fast computation of inner products between a signal and Gabor atoms}
\usage{
gabor_projection_fft(block, signal)
}
\arguments{
\item{block}{See the vignette for a description of the structure of blocks.}

\item{signal}{A numeric vector, matrix, or data frame representing the signal(s)
to be analyzed. Each column is treated as a separate channel.}
}
\value{
A list containing two matrices computed from windowed FFT segments of the signal:
\item{proj_mod_mtx}{Magnitudes of selected Gabor atom inner products
(absolute values of projection coefficients).}
\item{fft_bin_mtx}{Complex Fourier coefficients used to compute inner products with Gabor atoms.}
}
\description{
This function computes inner products between a windowed signal and a set of Gabor
atoms using FFT-based frequency-domain operations. Instead of explicitly constructing
and shifting atoms in the time domain, it extracts selected Fourier coefficients
corresponding to Gabor frequencies. The resulting values provide both complex
projection coefficients and their magnitudes.
}
\note{
This function is primarily intended for internal use by \code{topk_atoms()},
but it is exported to support advanced experiments and methodological testing.
}
\examples{
signal <- as.matrix(rnorm(256))
sampling_frequency <- 256
duration <- 1

xml_file <- system.file("extdata", "one_block.xml", package = "MatchingPursuit")
block <- read_gabor_dict(xml_file, sampling_frequency, duration, verbose = TRUE)
my_list <- gabor_projection_fft(block, signal)

pmm <- my_list$proj_mod_mtx
scm <- my_list$fft_bin_mtx

head(scm)
head(pmm)

# Of course it gives 'pmm'
head(Mod(scm))

}
