% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.empi.R
\name{locate.empi}
\alias{locate.empi}
\title{Get required external software localization}
\usage{
locate.empi()
}
\value{
List with URL of the \emph{empi} binaries and zip file name.
}
\description{
Returns \emph{Enhanced Matching Pursuit Implementation} \url{https://github.com/develancer/empi}
binary locations for the following operation systems: Windows, Linux, MacOS-x64, MacOS-arm64.
}
\examples{
locate.empi()

}
