% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MatchingPursuit}
\alias{MatchingPursuit-package}
\alias{MatchingPursuit}
\title{Sparse Time-Series Decomposition Using Matching Pursuit and Orthogonal Matching Pursuit}
\description{
Tools for analyzing and decomposing time_series data using the
\strong{Matching Pursuit (MP)} algorithm, a greedy signal decomposition
technique that represents complex signals as a linear combination of simpler
functions (called atoms) selected from a redundant dictionary. Support
for the \strong{Orthogonal Matching Pursuit (OMP)} variant of the classical MP
algorithm is also provided.
}
\details{
Both the MP and OMP algorithms only support Gabor atoms. However, both algorithms
are much more general. They can handle any atom dictionary, as long as we can
compute the dot products of the signal and the atoms. Gabor atoms are particularly
popular because they well implement the time-frequency tradeoff implied by the
Heisenberg Uncertainty Principle and describe many natural signals.

In addition to generic time-series data, the package supports direct loading of
data stored in \strong{EDF/EDF(+)} and \strong{WFDB} (WaveForm DataBase) formats.
These formats are widely used for physiological signals such as EEG and ECG recordings.
Support for EDF/EDF(+) and WFDB import facilitates the analysis of biomedical signals.

The package requires installation of an external program,
\strong{Enhanced Matching Pursuit Implementation (EMPI)}.
This tool implements the Matching Pursuit algorithm developed by
\strong{Piotr T. Różański} and is available at
\url{https://github.com/develancer/empi}

The package also provides a pure R implementation of the \strong{MP} algorithm.
It is primarily intended for reference and educational purposes and is
slower and less numerically precise than the \strong{EMPI} implementation.

Example datasets available via the \code{system.file()} function:

 \itemize{
   \item \code{EEG.edf}
     \itemize{
       \item 19 EEG channels + 1 EDF_Annotations channel
       \item sampling frequency: 256 Hz, signal length: 10 sec.
       \item channel names: Fp1, Fp2, F3, F4, F7, F8, Fz, C3, C4, Cz, T3, T5, T4, T6, P3, P4, Pz, O1, O2, EDF_Annotations
     }
   \item \code{EEG_filter_resample_montage.db}, \code{EEG_filter_resample_montage.csv}, \code{EEG_filter_resample_montage.bin}
     \itemize{
       \item 18 EEG channels after application of the double-banana montage, resampling and filtering of the \code{EEG.edf} data
       \item sampling frequency: 256 Hz, signal length: 10 sec (64 Hz after resampling).
       \item channel names: Fp2_F4, F4_C4, C4_P4, P4_O2, Fp1_F3, F3_C3, C3_P3, P3_O1, Fp2_F8, F8_T4,
                            T4_T6, T6_O2, Fp1_F7, F7_T3, T3_T5, T5_O1, Fz_Cz, Cz_Pz
     }
   \item \code{sample1.csv}, \code{sample1.db}
     \itemize{
       \item 1 channel
       \item sampling frequency: 1024 Hz, signal length: 1 sec.
     }
   \item \code{sample2.csv}, \code{sample2.db}
     \itemize{
       \item 1 channel
       \item sampling frequency: 128 Hz, signal length: 10 sec.
     }
   \item \code{sample3.csv}, \code{sample3.db}
     \itemize{
       \item 3 channels (sum of four sinusoids, burst signal, sum of three Gabor atoms)
       \item sampling frequency: 128 Hz, signal length: 2 sec.
     }
   \item \code{00001_lr.dat}, \code{00001_lr.hea}
     \itemize{
       \item Example ECG recording from \url{https://physionet.org/content/ptb-xl/1.0.3/}
       \item 12 ECG leads, 10 sec, 16-bit integer format
       \item standard lead names: I, II, III, aVR, aVL, aVF, V1–V6
     }
   \item \code{sample1_dict.xml}, \code{sample2_dict.xml}, \code{sample3_dict.xml}, \code{sample3_dict_EMPI.xml},
   \code{EEG_filter_resample_montage_dict.xml}, \code{one_block_dict.xml}, \code{00001_lr_dict.xml}
     \itemize{
       \item XML files describing a multiscale Gabor dictionary.
       \item such files can be generated from the EMPI program executed with the \code{--dictionary-output}
       option, which allows you to save (in XML format) data about the dictionary used.
       See the \code{read_dict()} function help page for examples and further details.
     }
 }

The first line of a \code{.csv} file contains two numbers: sampling rate in Hz (\code{freq})
and signal length in seconds (\code{sec}). The \code{read_csv_signals()} function verifies
whether the file contains exactly \code{round(freq * sec)} samples. The two numbers
must be separated by one or more whitespace characters.

Optionally, channel names may be specified in the second line of the \code{.csv} file.
In such cases, use \code{col_names_in_csv = TRUE} when calling \code{read_csv_signals()}.

Files with the \code{.db} extension are in \code{SQLite} format and are produced by
the \code{empi_execute()} function.
}
\section{Examples}{

A slightly longer demo script showing the most important functionality of the package:
\code{system.file("examples", "quickstart.R", package = "MatchingPursuit")}
}

\references{
Durka, P. J. (2007). \emph{Matching Pursuit and Unification in EEG Analysis}. Artech House, Engineering
in Medicine and Biology. Boston. ISBN: 978-1596932497

Elad, M. (2010). \emph{Sparse and Redundant Representations: From Theory to Applications in Signal
and Image Processing}. Springer. ISBN 978-1-4419-7010-7, \doi{10.1007/978-1-4419-7011-4}

Gramacki, A. & Kunik, M. (2025).
\emph{Deep learning epileptic seizure detection based on matching pursuit algorithm and its time-frequency
graphical representation}. International Journal of Applied Mathematics & Computer Science,
vol. 35, no. 4, pp. 617-630, \doi{10.61822/amcs-2025-0044}

Mallat, S. & Zhang, Z. (1993). \emph{Matching Pursuits with Time-Frequency Dictionaries}.
IEEE Transactions on Signal Processing, vol. 41, no. 12, pp. 3397-3415, \doi{10.1109/78.258082}

Pati, Y.C. & Rezaiifar, R. & Krishnaprasad, P.S. (1993). \emph{Orthogonal Matching Pursuit: Recursive
Function Approximation with Applications to Wavelet Decomposition}. Proceedings of the 27th Asilomar
Conference on Signals, Systems and Computers, vol. 1, pp. 40-44 \doi{10.1109/ACSSC.1993.342465}

Różański, P.T. (2024). \emph{empi: GPU-Accelerated Match ing Pursuit with Continuous Dictionaries}.
ACM Transactions on Mathematical Software, vol.50, no. 3, pp. 1-17, \doi{10.1145/3674832}
}
\author{
\strong{Maintainer}: Artur Gramacki \email{a.gramacki@gmail.com} (\href{https://orcid.org/0000-0002-1610-9743}{ORCID})

Other contributors:
\itemize{
  \item Jarosław Gramacki \email{j.gramacki@gmail.com} (\href{https://orcid.org/0000-0001-5032-1353}{ORCID}) [contributor]
  \item Piotr T. Różański \email{piotr@develancer.pl} (\href{https://orcid.org/0000-0002-0457-6731}{ORCID}) [contributor]
}

}
\keyword{internal}
