% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MatchingPursuit}
\alias{MatchingPursuit-package}
\alias{MatchingPursuit}
\title{Processing Time Series Data Using the Matching Pursuit Algorithm}
\description{
Tools for analyzing and decomposing time-series data using the
\strong{Matching Pursuit (MP)} algorithm, a greedy signal decomposition technique that represents
complex signals as a linear combination of simpler functions (called atoms) selected from a
redundant dictionary.
}
\details{
In addition to generic time-series data, the package supports direct loading of
data stored in EDF/EDF(+) and WFDB (WaveForm DataBase) formats. These formats are widely used for
physiological signals such as EEG and ECG recordings. Support for EDF/EDF(+) and WFDB
import facilitates the analysis of biomedical signals.

The package requires installation of an external program,
\strong{Enhanced Matching Pursuit Implementation} (EMPI).
This tool implements the Matching Pursuit algorithm developed by
\strong{Piotr T. Różański} and is available at
\url{https://github.com/develancer/empi}

Example datasets available via the \code{system.file()} function:

 \itemize{
   \item \code{EEG.edf}, \code{EEG.db}, \code{EEG.csv}
     \itemize{
       \item 19 EEG channels + 1 EDF Annotations channel
       \item sampling frequency: 256 Hz, signal length: 10 sec.
       \item channel names: Fp1, Fp2, F3, F4, F7, F8, Fz, C3, C4, Cz, T3, T5, T4, T6, P3, P4, Pz, O1, O2, EDF_Annotations
     }
   \item \code{EEG_bipolar_filtered.db}, \code{EEG_bipolar_filtered.db.csv}
     \itemize{
       \item 18 EEG channels after double banana montage and filtering of the data from \code{EEG.edf}
       \item sampling frequency: 256 Hz, signal length: 10 sec.
       \item channel names: Fp2_F4, F4_C4, C4_P4, P4_O2, Fp1_F3, F3_C3, C3_P3, P3_O1, Fp2_F8, F8_T4,
                            T4_T6, T6_O2, Fp1_F7, F7_T3, T3_T5, T5_O1, Fz_Cz, Cz_Pz
     }
   \item \code{sample1.csv}, \code{sample1.db}
     \itemize{
       \item 1 channel
       \item sampling frequency: 1024 Hz, signal length: 1 sec.
     }
   \item \code{sample2.csv}, \code{sample2.db}
     \itemize{
       \item 1 channel
       \item sampling frequency: 128 Hz Hz, signal length: 10 sec.
     }
   \item \code{sample3.csv}, \code{sample3.db}
     \itemize{
       \item 3 channels (random numbers from 0 to 1 in each channel)
       \item sampling frequency: 128 Hz Hz, signal length: 2 sec.
     }
   \item \code{00001_lr.dat}, \code{00001_lr.hea}
     \itemize{
       \item Example ECG recording from \url{https://physionet.org/content/ptb-xl/1.0.3/}
       \item 12 ECG leads, 10 sec, 16-bit integer format
       \item standard lead names: I, II, III, aVR, aVL, aVF, V1–V6
     }
}
The first line of a \code{.csv} file contains two numbers: sampling rate in Hz (\code{freq})
and signal length in seconds (\code{sec}). The \code{read.csv.signals()} function verifies
whether the file contains exactly \code{round(freq * sec)} samples. The two numbers
must be separated by one or more whitespace characters.

Optionally, channel names may be specified in the second line of the \code{.csv} file.
In such cases, use \code{col.names.in.csv = TRUE} when calling \code{read.csv.signals()}.

Files with the \code{.db} extension are in \code{SQLite} format and are produced by
the \code{empi.execute()} function.
}
\examples{
if (interactive()) {
  ###############################################################################
  # Clear the cache directory, install the 'empi' program,
  # and verify that the installation completed successfully.
  ###############################################################################
  clear.cache()
  empi.install()
  empi.check()

  ###############################################################################
  # Workflow 1. Data stored in a CSV file.
  ###############################################################################
  # STEP 1 - Read sample data.
  file.sample1.csv <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
  signal.sample1.csv <- read.csv.signals(file.sample1.csv, col.names = "ch1")

  # STEP 2 - Run the MP algorithm.
  empi.class <- empi.execute(signal = signal.sample1.csv)

  # STEP 3 - Plot the time-frequency map based on MP atoms.
  # plot.empi() is the S3 method for the generic plot() function.
  # It requires an object of class empi, created with empi.execute().
  plot(empi.class)

  # STEP 3.bis — This call gives the same result as plot(empi.class).
  # Uncomment to check.
  # out <- empi2tf(
  #  x = empi.class,
  #  channel = 1,
  #  increase.factor = 8,
  #  out.mode = "plot"
  #)

  ###############################################################################
  # Workflow 2. Data stored in an EDF file (EEG data).
  ###############################################################################
  # STEP 1 - Read a sample EEG file.
  file.EEG.edf <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
  out.EEG <- read.edf.signals(file.EEG.edf)
  signal.EEG <- out.EEG$signal
  sampling.rate <- out.EEG$sampling.rate

  # STEP 2 - Create the EEG montage.
  # This step is typically required during EEG processing.
  # Signal pairs used to construct the bipolar montage
  # (the so-called "double banana" montage).
  pairs <- list(
    c("Fp2", "F4"), c("F4", "C4"), c("C4", "P4"), c("P4", "O2"), c("Fp1", "F3"), c("F3", "C3"),
    c("C3", "P3"), c("P3", "O1"), c("Fp2", "F8"), c("F8", "T4"), c("T4", "T6"), c("T6", "O2"),
    c("Fp1", "F7"), c("F7", "T3"), c("T3", "T5"), c("T5", "O1"), c("Fz", "Cz"), c("Cz", "Pz")
  )

  bip.montage <- eeg.montage(out.EEG, montage.type = c("bipolar"), bipolar.pairs = pairs)

  # STEP 3 - Filter the data.
  # This step is typically required during EEG processing.
  # Define commonly used EEG filtering parameters.
  fc <- filters.coeff(
    fs = sampling.rate,
    notch = c(49, 51),
    lowpass = 40,
    highpass = 1,
  )

  sig.filt <- bip.montage$signal

  for (m in 1:ncol(sig.filt)) {
    # 50Hz notch filter
    sig.filt[, m] = signal::filtfilt(fc$notch, sig.filt[, m])
    # Low pass IIR Butterworth
    sig.filt[, m] = signal::filtfilt(fc$lowpass, sig.filt[, m])
    # High pass IIR Butterwoth
    sig.filt[, m] = signal::filtfilt(fc$highpass, sig.filt[, m])
  }

  signal.EEG.montage.filt <- list(sig.filt, sampling.rate)
  names(signal.EEG.montage.filt) <- c("signal", "sampling.rate")

  # STEP 4 - Run the MP algorithm.
  empi.class <- empi.execute(signal = signal.EEG.montage.filt)

  # STEP 5 - Plot the time-frequency map based on MP atoms.
  plot(empi.class, channel = 2)

  # STEP 6 - Plot EEG signals after applying the bipolar montage
  # and the filtering procedure.
  temp <- bip.montage
  temp$signal <- sig.filt

  plot(
    x = temp,
    begin = 0,
    end = 10,
    panel.height = NULL,
    rainbow = FALSE,
    bg.colour = "white",
    txt.col = "blue",
    zero.line = TRUE,
    main = "EEG.edf file after banana montage and after filtering"
  )

  ###############################################################################
  # Workflow 3. Data stored in WFDB (WaveForm DataBase) format.
  ###############################################################################
  # STEP 1 - Read sample data.
  file.00001_lr.hea <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")
  out.ecg <- read.ecg.signals(file.00001_lr.hea)

  # Create a list compatible with the empi.execute() function.
  signal.ecg <- list(
    signal = data.frame(out.ecg$signal),
    sampling.rate = out.ecg$sampling.rate
  )

  # STEP 2 - Run the MP algorithm.
  empi.class <- empi.execute(signal = signal.ecg)

  # STEP 3 - Plot the time-frequency map based on MP atoms.
  plot(empi.class)

  # STEP 4 - Display ECG signals using a layout corresponding
  # to standard ECG paper.
  # A typical ECG paper layout was used, with a small grid
  # of 0.04 s × 0.1 mV and a large grid of 0.20 s × 0.5 mV.
  plot(
    x = out.ecg,
    begin = 0,
    end = 10,
    panel.height = 1,
    zero.line = FALSE,
    small.squares = TRUE
  )

  ###############################################################################
  # Save empi.execute() results to a SQLite file.
  ###############################################################################
  # Run the MP algorithm and save the results to a SQLite file.
  empi.class <- empi.execute(
    signal = signal.sample1.csv,
    write.to.file = TRUE,
    path = ".",
    file.name = "sample1.db")

  # Read atom parameters.
  atoms <- atom.params(db.file = "sample1.db")
  print(atoms)

  ###############################################################################
  # Some useful options in empi2tf() function.
  ###############################################################################
  # Save the time-frequency map as a PNG file
  # with the specified dimensions (in pixels).
  out <- empi2tf(
    x = empi.class,
    channel = 1,
    increase.factor = 8,
    out.mode = "file",
    path = ".",
    file.name = "sample1.png",
    size = c(512, 512)
  )

  # Save the time-frequency map as an RData file
  # containing a matrix with the specified dimensions.
  out <- empi2tf(
    x = empi.class,
    channel = 1,
    out.mode = "RData",
    path = ".",
    file.name = "sample1.RData",
    size = c(128, 128)
  )

  # Unlike plot.empi(), empi2tf() can also accept
  # a path to a SQLite file created by empi.execute().
  file.sample1.db <- system.file("extdata", "sample1.db", package = "MatchingPursuit")

  out <- empi2tf(
    x = file.sample1.db,
    channel = 1,
    increase.factor = 8,
    out.mode = "plot"
  )

}

}
\references{
Durka, P. J. (2007). \emph{Matching Pursuit and Unification in EEG Analysis}. Artech House, Engineering
in Medicine and Biology. Boston. ISBN: 978-1596932497

Elad, M. (2010). \emph{Sparse and Redundant Representations: From Theory to Applications in Signal
and Image Processing}. Springer. ISBN 978-1-4419-7010-7, \doi{10.1007/978-1-4419-7011-4}

Gramacki, A. & Kunik, M. (2025).
\emph{Deep learning epileptic seizure detection based on matching pursuit algorithm and its time-frequency
graphical representation}. International Journal of Applied Mathematics & Computer Science,
vol. 35, no. 4, pp. 617-630, \doi{10.61822/amcs-2025-0044}

Mallat, S. & Zhang, Z. (1993). \emph{Matching Pursuits with Time-Frequency Dictionaries}.
IEEE Transactions on Signal Processing, vol. 41, no. 12, pp. 3397-3415, \doi{10.1109/78.258082}

Pati, Y.C. & Rezaiifar, R. & Krishnaprasad, P.S. (1993). \emph{Orthogonal Matching Pursuit: Recursive
Function Approximation with Applications to Wavelet Decomposition}. Proceedings of the 27th Asilomar
Conference on Signals, Systems and Computers, vol. 1, pp. 40-44 \doi{10.1109/ACSSC.1993.342465}

Różański, P.T. (2024). \emph{empi: GPU-Accelerated Match ing Pursuit with Continuous Dictionaries}.
ACM Transactions on Mathematical Software, vol.50, no. 3, pp. 1-17, \doi{10.1145/3674832}
}
\author{
\strong{Maintainer}: Artur Gramacki \email{a.gramacki@gmail.com} (\href{https://orcid.org/0000-0002-1610-9743}{ORCID})

Other contributors:
\itemize{
  \item Jarosław Gramacki \email{j.gramacki@gmail.com} (\href{https://orcid.org/0000-0001-5032-1353}{ORCID}) [contributor]
  \item Piotr T. Różański \email{piotr@develancer.pl} (\href{https://orcid.org/0000-0002-0457-6731}{ORCID}) [contributor]
}

}
\keyword{internal}
