% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MatchingPursuit}
\alias{MatchingPursuit-package}
\alias{MatchingPursuit}
\title{Using the Matching Pursuit algorithm to analyze time series data.}
\description{
Dłuższy opis pakietu. Możesz tu opisać:
- do czego służy pakiet
- główne funkcje
- przykładowy workflow
}
\details{
todo
}
\references{
Mallat, S. & Zhang, Z. (1993). \emph{Matching Pursuits with Time-Frequency Dictionaries}.
IEEE Transactions on Signal Processing, vol. 41, no. 12, pp. 3397-3415, \url{https://doi.org/10.1109/78.258082}

Tropp, J.A. & Gilbert, A.C. (2007). \emph{Signal Recovery From Random Measurements via Orthogonal Matching}.
IEEE Transactions on Information Theory, vol. 53, no. 12, pp. 4655-4666, \url{https://doi.org/10.1109/TIT.2007.909108}

Różański, P.T. (2024). \emph{empi: GPU-Accelerated Matching Pursuit with Continuous Dictionaries}.
ACM Transactions on Mathematical Software, vol.50, issue = 3, pp. 1-17, \url{https://doi.org/10.1145/3674832}

 Gramacki, A. & Kunik, M. (2025).
 \emph{Deep learning epileptic seizure detection based on matching pursuit algorithm and its time-frequency graphical representation}. International Journal of Applied Mathematics & Computer Science, vol. 35, no. 4, pp. 617-630, \url{https://doi.org//10.61822/amcs-2025-0044}
}
\author{
\strong{Maintainer}: Artur Gramacki \email{a.gramacki@gmail.com} (\href{https://orcid.org/0000-0002-1610-9743}{ORCID})

Other contributors:
\itemize{
  \item Jarosław Gramacki \email{a.gramacki@gmail.com} (\href{https://orcid.org/0000-0001-5032-1353}{ORCID}) [contributor]
}

}
\keyword{internal}
