% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MatchingPursuit}
\alias{MatchingPursuit-package}
\alias{MatchingPursuit}
\title{Processing Time Series Data Using the Matching Pursuit Algorithm}
\description{
Tools for analyzing and decomposing time-series data using the
\strong{Matching Pursuit (MP)} algorithm, a greedy signal decomposition technique that represents
complex signals as a linear combination of simpler functions (called atoms) selected from a
redundant dictionary.
}
\details{
In addition to working with generic time-series data, the package also supports direct loading of
data stored in EDF and EDF(+) files. These formats are widely used for storing physiological signals
such as EEG, EMG, or ECG recordings. By enabling the import of EDF and EDF(+) files, the package
facilitates the analysis of biomedical signals. The package requires the installation of an external
program, \strong{Enhanced Matching Pursuit Implementation} (EMPI). This tool implements the Matching
Pursuit algorithm developed by \strong{Piotr T. Różański} and is available
at \url{https://github.com/develancer/empi}

Example datasets available through the \code{system.file()} function are:

 \itemize{
   \item \code{EEG.edf}, \code{EEG.db}, \code{EEG.csv}
     \itemize{
       \item 19 EEG channels + 1 EDF Annotations channel
       \item sampling frequency: 256 Hz, signal length: 10 sec.
       \item channel names: Fp1, Fp2, F3, F4, F7, F8, Fz, C3, C4, Cz, T3, T5, T4, T6, P3, P4, Pz, O1, O2, EDF_Annotations
     }
   \item \code{EEG_bipolar_filtered.db}, \code{EEG_bipolar_filtered.db.csv}
     \itemize{
       \item 18 EEG channels after double banana montage and filtering of the data from \code{EEG.edf}
       \item sampling frequency: 256 Hz, signal length: 10 sec.
       \item channel names: Fp2_F4, F4_C4, C4_P4, P4_O2, Fp1_F3, F3_C3, C3_P3, P3_O1, Fp2_F8, F8_T4,
                            T4_T6, T6_O2, Fp1_F7, F7_T3, T3_T5, T5_O1, Fz_Cz, Cz_Pz
     }
   \item \code{sample1.csv}, \code{sample1.db}
     \itemize{
       \item 1 channel
       \item sampling frequency: 1024 Hz, signal length: 1 sec.
     }
   \item \code{sample2.csv}, \code{sample2.db}
     \itemize{
       \item 1 channel
       \item sampling frequency: 128 Hz Hz, signal length: 10 sec.
     }
   \item \code{sample3.csv}, \code{sample3.db}
     \itemize{
       \item 3 channels (random numbers from 0 to 1 in each channel)
       \item sampling frequency: 128 Hz Hz, signal length: 2 sec.
     }
}
The first line of the \code{.csv} file contains two numbers: the sampling rate in Hz (\code{freq})
and the signal length in seconds (\code{sec}). \code{read.csv.signals} function checks
whether the file actually contains \code{round(freq*sec)} samples. The two numbers
must by separated by one or more whitespace characters.

Optionally, channel names can be specified in the second line of the \code{.csv} file.
For such files, when reading their contents using the \code{read.csv.signals()} function,
use the \code{col.names.in.csv=TRUE} parameter.

Files with the \code{db} extension are in the \code{SQLite} format and are the result of
running the \code{empi.execute()} function.
}
\examples{
if (interactive()) {
  ###############################################################################
  # Clear the cache directory, install the 'empi' program and
  # check if the 'empi' program is installed.
  ###############################################################################
  clear.cache()
  empi.install()
  empi.check()

  ###############################################################################
  # Workflow 1. Data are stored in a CSV file.
  ###############################################################################
  # STEP 1 - read sample data.
  file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
  signal <- read.csv.signals(file, col.names = "ch1")

  # STEP 2 - execute the MP algorithm.
  empi.class <- empi.execute(signal = signal)

  # STEP 3 - plot a time-frequency map based on MP atoms.
  plot(empi.class)

  ###############################################################################
  # Workflow 2. Data are stored in a EDF file (EEG results).
  ###############################################################################
  # STEP 1 - reading a sample EEG file.
  file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
  out.EEG <- read.edf.signals(file)
  signal.EEG <- out.EEG$signals
  sampling.rate <- out.EEG$sampling.rate

  # STEP 2 - create the EEG montage.
  # This step is usually required while processing EEG data.
  # Pairs of signals for bipolar montage (so called "double banana").
  pairs <- list(
    c("Fp2", "F4"), c("F4", "C4"), c("C4", "P4"), c("P4", "O2"), c("Fp1", "F3"), c("F3", "C3"),
    c("C3", "P3"), c("P3", "O1"), c("Fp2", "F8"), c("F8", "T4"), c("T4", "T6"), c("T6", "O2"),
    c("Fp1", "F7"), c("F7", "T3"), c("T3", "T5"), c("T5", "O1"), c("Fz", "Cz"), c("Cz", "Pz")
  )

  bip.montage <- eeg.montage(signal.EEG, montage.type = c("bipolar"), bipolar.pairs = pairs)

  # STEP 3 - filter the data
  # This step is usually required while processing EEG data.
  # Filter parameters that will be used (quite typical in filtering EEG signals).
  fc <- filters.coeff(
    fs = sampling.rate,
    notch = c(49, 51),
    lowpass = 40,
    highpass = 1,
  )

  bip.montage.filt <- bip.montage

  for (m in 1:ncol(bip.montage)) {
    # 50Hz notch filter
    bip.montage.filt[, m] = signal::filtfilt(fc$notch, bip.montage[, m])
    # Low pass IIR Butterworth
    bip.montage.filt[, m] = signal::filtfilt(fc$lowpass, bip.montage.filt[, m])
    # High pass IIR Butterwoth
    bip.montage.filt[, m] = signal::filtfilt(fc$highpass, bip.montage.filt[, m])
  }

  signal.EEG.montage.filt <- list(bip.montage.filt, sampling.rate)
  names(signal.EEG.montage.filt) <- c("signal", "sampling.rate")

  # STEP 4 - execute the MP algorithm.
  empi.class <- empi.execute(signal = signal.EEG.montage.filt)

  # STEP 5 - plot a time-frequency map based on MP atoms.
  plot(empi.class, channel = 2)

  ###############################################################################
  # Saving the empi.execute() results to a SQLite file.
  ###############################################################################
  # Execute the MP algorithm and save the result to a SQLite file.
  empi.class <- empi.execute(
    signal = signal,
    write.to.file = TRUE,
    path = ".",
    file.name = "sample1.db")

  # Read the atoms parameters.
  atom.params(db.file = "sample1.db")

  ###############################################################################
  # empi2tf() function.
  ###############################################################################
  # This function call produces the same effect as plot(empi.class).
  out <- empi2tf(
    x = empi.class,
    channel = 1,
    increase.factor = 8,
    out.mode = "plot"
  )

  # Save the T-F map as a PNG file with the specified dimensions (in pixels).
  out <- empi2tf(
    x = empi.class,
    channel = 1,
    increase.factor = 8,
    out.mode = "file",
    path = ".",
    file.name = "sample1.png",
    size = c(512, 512)
  )

  # Save the T-F map as an RData file containing a matrix with the specified dimensions.
  out <- empi2tf(
    x = empi.class,
    channel = 1,
    out.mode = "RData",
    path = ".",
    file.name = "sample1.RData",
    size = c(128, 128)
  )
}

}
\references{
Durka, P. J. (2007). \emph{Matching Pursuit and Unification in EEG Analysis}. Artech House, Engineering
in Medicine and Biology. Boston. ISBN: 978-1596932497

Elad, M. (2010). \emph{Sparse and Redundant Representations: From Theory to Applications in Signal
and Image Processing}. Springer. ISBN 978-1-4419-7010-7, \doi{10.1007/978-1-4419-7011-4}

Gramacki, A. & Kunik, M. (2025).
\emph{Deep learning epileptic seizure detection based on matching pursuit algorithm and its time-frequency
graphical representation}. International Journal of Applied Mathematics & Computer Science,
vol. 35, no. 4, pp. 617-630, \doi{10.61822/amcs-2025-0044}

Mallat, S. & Zhang, Z. (1993). \emph{Matching Pursuits with Time-Frequency Dictionaries}.
IEEE Transactions on Signal Processing, vol. 41, no. 12, pp. 3397-3415, \doi{10.1109/78.258082}

Pati, Y.C. & Rezaiifar, R. & Krishnaprasad, P.S. (1993). \emph{Orthogonal Matching Pursuit: Recursive
Function Approximation with Applications to Wavelet Decomposition}. Proceedings of the 27th Asilomar
Conference on Signals, Systems and Computers, vol. 1, pp. 40-44 \doi{10.1109/ACSSC.1993.342465}

Różański, P.T. (2024). \emph{empi: GPU-Accelerated Match ing Pursuit with Continuous Dictionaries}.
ACM Transactions on Mathematical Software, vol.50, no. 3, pp. 1-17, \doi{10.1145/3674832}
}
\author{
\strong{Maintainer}: Artur Gramacki \email{a.gramacki@gmail.com} (\href{https://orcid.org/0000-0002-1610-9743}{ORCID})

Other contributors:
\itemize{
  \item Jarosław Gramacki \email{j.gramacki@gmail.com} (\href{https://orcid.org/0000-0001-5032-1353}{ORCID}) [contributor]
  \item Piotr T. Różański \email{piotr@develancer.pl} (\href{https://orcid.org/0000-0002-0457-6731}{ORCID}) [contributor]
}

}
\keyword{internal}
