% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ecg.R
\name{plot.ecg}
\alias{plot.ecg}
\title{The function displays ECG signals in a layout corresponding to standard paper ECG printouts}
\usage{
\method{plot}{ecg}(
  x,
  begin,
  end,
  panel.height = 3,
  small.squares = TRUE,
  zero.line = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{ecg} (from \code{read.ecg.signals()}).}

\item{begin}{Timepoint as representing seconds to start at.}

\item{end}{Timepoint as representing seconds to stop at.}

\item{panel.height}{Number of large squares to display (according to standard ECG paper):
\itemize{
    \item small grid: 0.04 sec. x 0.1 mV
    \item large grid: 0.20 sec. x 0.5 mV
}}

\item{small.squares}{If \code{TRUE}, also small grid will be displayed.}

\item{zero.line}{If \code{TRUE}, the horizontal line representing
\code{0 mV} will be displayed.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize an EKG graph.
}
\description{
A typical ECG paper layout was used, with a small grid of 0.04 s × 0.1 mV and a
large grid of 0.20 s × 0.5 mV.
}
\examples{
# ECG data comes from https://physionet.org/content/ptb-xl/1.0.3/
file <- system.file("extdata", "00001_lr.hea", package = "MatchingPursuit")
dir <- dirname(file)
name <- tools::file_path_sans_ext(basename(file))

out <- read.ecg.signals(file)

plot(
  x = out,
  begin = 0,
  end = 10,
  panel.height = 2,
  zero.line = FALSE,
  small.squares = TRUE
)

}
