% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.edf.R
\name{plot.edf}
\alias{plot.edf}
\title{The function displays EEG signals}
\usage{
\method{plot}{edf}(
  x,
  begin,
  end,
  panel.height = NULL,
  rainbow = TRUE,
  bg.colour = "black",
  txt.col = "white",
  zero.line = TRUE,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class \code{edf} (from \code{read.edf.signals()}).}

\item{begin}{Timepoint as representing seconds to start at.}

\item{end}{Timepoint as representing seconds to stop at.}

\item{panel.height}{Controls the vertical distance between individual signals.
If \code{NULL}, its value is set automatically so that all signals are clearly
visible and do not overlap}

\item{rainbow}{if \code{TRUE}, individual channels will be drawn in different colors}

\item{bg.colour}{Background color}

\item{txt.col}{Colour of text elements (X and Y axis, title).}

\item{zero.line}{If \code{TRUE}, the horizontal line representing
\code{0 mV} will be displayed.}

\item{main}{The text shown as the plot title.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize an EEG graph.
}
\description{
Signals are displayed one below the other. They can be displayed in different
colors for easier viewing.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out  <- read.edf.signals(file, resampling = FALSE)

plot(
  x = out,
  begin = 0,
  end = 10,
  panel.height = NULL,
  rainbow = TRUE,
  bg.colour = "black",
  txt.col = "white",
  zero.line = TRUE,
  main = "EEG.edf file after banana montage and after filtering"
)

plot(
  x = out,
  begin = 0,
  end = 10,
  panel.height = NULL,
  rainbow = FALSE,
  bg.colour = "white",
  txt.col = "black",
  zero.line = TRUE,
  main = "EEG.edf file after banana montage and after filtering"
)
}
