% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.edf.R
\name{plot.edf}
\alias{plot.edf}
\title{The function displays EEG signals}
\usage{
\method{plot}{edf}(
  x,
  begin,
  end,
  panel_height = NULL,
  rainbow = TRUE,
  bg_colour = "black",
  txt_col = "white",
  zero_line = TRUE,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class \code{edf} (from \code{read_edf_ signals()}).}

\item{begin}{Time point (in seconds) at which to start plotting.}

\item{end}{Time point (in seconds) at which to stop plotting.}

\item{panel_height}{Controls the vertical spacing between individual signals.
If \code{NULL}, the value is chosen automatically so that all signals are clearly
visible and do not overlap.}

\item{rainbow}{If \code{TRUE}, individual channels are drawn in different colours.}

\item{bg_colour}{Background colour.}

\item{txt_col}{Colour of text elements (axis labels and title).}

\item{zero_line}{If \code{TRUE}, a horizontal line representing \code{0 mV} is displayed.}

\item{main}{The text shown as the plot title.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize an EEG graph.
}
\description{
Signals are displayed one below another and may be shown in different
colours for improved readability.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out  <- read_edf_signals(file, resampling = FALSE)

plot(
  x = out,
  begin = 0,
  end = 10,
  panel_height = NULL,
  rainbow = TRUE,
  bg_colour = "black",
  txt_col = "white",
  zero_line = TRUE,
  main = "EEG signals stored in the EEG.edf file"
)

plot(
  x = out,
  begin = 0,
  end = 10,
  panel_height = NULL,
  rainbow = FALSE,
  bg_colour = "white",
  txt_col = "black",
  zero_line = TRUE,
  main = "EEG signals stored in the EEG.edf file"
)
}
