% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.empi.R
\name{plot.empi}
\alias{plot.empi}
\title{plots T-F map to visualize the empi decomposition}
\usage{
\method{plot}{empi}(
  x,
  channel = 1,
  mode = "sqrt",
  freq.divide = NULL,
  increase.factor = 8,
  shortening.factor.x = 2,
  shortening.factor.y = 2,
  display.crosses = TRUE,
  display.atom.numbers = FALSE,
  display.grid = FALSE,
  crosses.color = "white",
  palette = "my custom palette",
  plot.signals = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{empi} object created after executing the \code{empi.execute()} function.
In this case, the \code{db.file} parameter must be \code{NULL}.}

\item{channel}{Channel from the SQLite file to process.}

\item{mode}{\code{"sqrt"}, \code{"log"}, or \code{"linear"}. It determines the intensity
with which the so-called blobs are displayed on the T-F map.}

\item{freq.divide}{Specifies how many times the displayed frequency in the T-F map
should be decreased. At high sampling rates and when we use a low-pass filter with
a typical cut-off frequency much lower than the sampling frequency, a large part of
the T-F map does not contain any blobs. If the sampling frequency is \code{f},
the maximum frequency in the T-F map will be \code{ceiling(f/2/freq.divide)}
(f/2 is the Nyquist rule). If \code{NULL}, it is determined based on the atom
with the highest frequency \code{fmax} according to \code{freq.divide=(f/2)/fmax}.}

\item{increase.factor}{Factor of increasing the number of pixels in the f-axis, the most
sensible are non-negative integers (e.g. 2, 4, 5, 8).}

\item{shortening.factor.x}{Usually, for better visualization of atoms, a value of 2 will
be appropriate.}

\item{shortening.factor.y}{Usually, for better visualization of atoms, a value of 2 will
be appropriate.}

\item{display.crosses}{Whether small crosses should be displayed in the canters of atoms.}

\item{display.atom.numbers}{Whether atom numbers should be displayed in the canters of atoms.}

\item{display.grid}{Whether to draw grid lines.}

\item{crosses.color}{Colour of small crosses.}

\item{palette}{Palette from the list returned by \code{hcl.pals()} function or the string
\code{"my custom palette"}.}

\item{plot.signals}{Whether the original and reconstructed signals should also be displayed.}

\item{...}{Currently ignored. Required for compatibility with the generic \code{plot()}.}
}
\value{
No return value, called to visualize the empi decomposition.
}
\description{
The function is in fact a wrapper for \code{empi2tf()} function where \code{out.mode = "plot"}.
}
\examples{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
signal <- read.csv.signals(file, col.names = "ch1")

# Execute the MP algorithm.
empi.out <- empi.execute(signal = signal)

# Plot a time-frequency map based on MP atoms.
plot(empi.out)

}
