% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csv.signals.R
\name{read.csv.signals}
\alias{read.csv.signals}
\title{Reads and validates a CSV file structure}
\usage{
read.csv.signals(file, col.names = NULL, col.names.in.csv = FALSE)
}
\arguments{
\item{file}{File to be read and checked. The first line of the file must contain two numbers:
the sampling rate in Hz (\code{freq}) and the signal length in seconds (\code{sec}).
The function verifies whether the file contains exactly \code{round(freq * sec)} samples.
The two numbers must be separated by one or more whitespace characters.}

\item{col.names}{Optional character vector of column names. If not specified, default names are created.}

\item{col.names.in.csv}{iLogical value. If \code{TRUE}, the second line of the file is assumed
to contain column names.}
}
\value{
A list containing:
\itemize{
  \item a data frame (rows = samples, columns = channels),
  \item sampling rate.
}
}
\description{
Reads and validates a CSV file structure
}
\examples{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")

# The first line of the file must contain two numbers:
# a) the sampling rate in Hz
# b) the signal length in seconds
out <- read.csv(file, header = FALSE)
head(out)

signal <- read.csv.signals(file, col.names = "signal_1")
head(signal$signal)
signal$sampling.rate

file <- system.file("extdata", "sample3.csv", package = "MatchingPursuit")
signal <- read.csv.signals(file, col.names = c("signal_1", "signal_2", "signal_3"))
head(signal$signal)
signal$sampling.rate

# Now, the csv file contains signal names in the second line
file <- system.file("extdata", "EEG.csv", package = "MatchingPursuit")
signal <- read.csv.signals(file, col.names.in.csv = TRUE)
head(signal$signal)
signal$sampling.rate

# Now, the csv file contains signal names in the second line
# The data here is the same as in the EEG.csv file, but after performing
# 'double banana' montages and after applying filtering.
file <- system.file("extdata", "EEG_bipolar_filtered.csv", package = "MatchingPursuit")
signal <- read.csv.signals(file, col.names.in.csv = TRUE)
head(signal$signal)
signal$sampling.rate

}
