% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csv.signals.R
\name{read.csv.signals}
\alias{read.csv.signals}
\title{Checks if the csv file has the correct structure}
\usage{
read.csv.signals(file, col.names = NULL)
}
\arguments{
\item{file}{File to be checked. The first line of the file must contain two numbers:
the sampling rate in Hz (\code{f}) and the signal length in seconds (\code{s}).
This checks whether the file actually contains \code{f*s} samples.}

\item{col.names}{Vector with column names. If not specified, default names will be created.}
}
\value{
Data frame where rows = samples for all channels, columns = channels.
}
\description{
Checks if the csv file has the correct structure
}
\examples{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")
signal <- read.csv.signals(file, col.names = "signal_1")
head(signal)

}
