% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.edf.params.R
\name{read.edf.params}
\alias{read.edf.params}
\title{Reads a selected EDF and EDF+ file and returns signals parameters.}
\usage{
read.edf.params(file)
}
\arguments{
\item{file}{The full path to the EDF/EDF+ file to be read.}
}
\value{
A data frame is returned.
}
\description{
The function reads a selected EDF(+) file and returns selected signals parameters
(channel names, frequency of each channel, number of samples in each channel
and the length of each channel in seconds).
}
\examples{
file <- system.file("extdata", "EEG_data_10sec.edf", package = "MatchingPursuit")
read.edf.params(file)

}
