% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.edf.signals.R
\name{read.edf.signals}
\alias{read.edf.signals}
\title{Reads a selected EDF or EDF+ file and returns all signals data}
\usage{
read.edf.signals(
  file,
  resampling = FALSE,
  f.new = NULL,
  from = NULL,
  to = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file}{The path to the EDF / EDF+ file to be read.}

\item{resampling}{If \code{TRUE} the frequency of all signals will be
upsampling or downsampling, depending on the actual sampling rate of subsequent channel.}

\item{f.new}{A new frequency (used for upsampling or downsampling).}

\item{from}{Loading a signal \code{from} (given as a second).}

\item{to}{Loading a signal \code{to} (given as a second).}

\item{verbose}{Flag to print out progress information.}
}
\value{
An object of class \code{edf}. The returned value is a list containing:
1) data frame with all signals stored in the given \code{edf} file,
2) sampling rate of the data after possible resampling (upsampled or downsampled),
3) time stamps of the data after possible resampling (upsampled or downsampled).
4) signal names
}
\description{
The function reads a selected EDF or EDF+ file.
Also resampling can be done (upsampling or downsampling).
}
\details{
If \code{resampling=TRUE}, the frequency of all signals will be upsampled or downsampled,
depending on the actual sampling rate of the individual channels and the set value of the
\code{f.new} parameter. The EDF standard assumes that each channel can be sampled at a different
rate. Therefore, it may happen that some channels are upsampled and others are downsampled. The
function does not provide the functionality to independently change the sampling rate for each channel.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out1  <- read.edf.signals(file, resampling = FALSE)

lapply(out1, class)
out1$sampling.rate

out2 <- read.edf.signals(file, resampling = TRUE, f.new = 128, verbose = TRUE)

lapply(out2, class)
out2$sampling.rate

}
