% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.edf.signals.R
\name{read.edf.signals}
\alias{read.edf.signals}
\title{Reads a selected EDF or EDF+ file and returns signal data}
\usage{
read.edf.signals(
  file,
  resampling = FALSE,
  f.new = NULL,
  from = NULL,
  to = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file}{Path to the EDF / EDF+ file to be read.}

\item{resampling}{If \code{TRUE}, all signals are resampled
(either upsampled or downsampled), depending on the original sampling rates of the channels.}

\item{f.new}{Target sampling frequency used for upsampling or downsampling.}

\item{from}{Starting time of the signal to be loaded (in seconds).}

\item{to}{Ending time of the signal to be loaded (in seconds).}

\item{verbose}{Logical flag indicating whether progress information should be printed.}
}
\value{
An object of class \code{edf}, a list containing:
1) a data frame with all signals stored in the EDF file,
2) sampling rate after optional resampling,
3) time stamps after optional resampling,
4) signal names.
}
\description{
The function reads a selected EDF or EDF+ file. Optionally, resampling can be performed
(upsampling or downsampling).
}
\details{
If \code{resampling = TRUE}, signals are resampled according to the target frequency
specified by \code{f.new}. Since the EDF standard allows different sampling rates per channel,
some channels may be upsampled while others are downsampled. The function does not support
independent resampling of individual channels.
}
\examples{
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out1  <- read.edf.signals(file, resampling = FALSE)

lapply(out1, class)
out1$sampling.rate

out2 <- read.edf.signals(file, resampling = TRUE, f.new = 128, verbose = TRUE)

lapply(out2, class)
out2$sampling.rate

}
