% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.edf.signals.R
\name{read.edf.signals}
\alias{read.edf.signals}
\title{Reads a selected EDF or EDF+ file and returns all signals data}
\usage{
read.edf.signals(
  file,
  resampling = TRUE,
  f.new = NULL,
  from = NULL,
  to = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file}{The full path to the EDF/EDF+ file to be read.}

\item{resampling}{Logical TRUE or FALSE. If TRUE the frequency of all signals will be
upsampling or downsampling, depending on the actual sampling rate of subsequent channel.}

\item{f.new}{A new frequency.}

\item{from}{Loading a signal \code{from} the given second.}

\item{to}{Loading a signal \code{to} the given second.}

\item{verbose}{Flag to print out progress information.}
}
\value{
A matrix \code{edf.mtx} with all signals data is returned and a list \code{edf} as a
result of executing \code{edf::read.edf()} function.
}
\description{
The function reads a selected EDF or EDF+ file and returns all signals data as a matrix.
Also resampling can be done (upsampling or downsampling).
}
\details{
If \code{resampling=TRUE}, the frequency of all signals will be upsampled or downsampled,
depending on the actual sampling rate of the individual channels and the set value of the
\code{f.new} parameter. The EDF standard assumes that each channel can be sampled at a different
rate. Therefore, it may happen that some channels are upsampled and others are downsampled. The
function does not provide the functionality to independently change the sampling rate for each channel.
}
\examples{
file <- system.file("extdata", "EEG_data_10sec.edf", package = "MatchingPursuit")
sigs1  <- read.edf.signals(file, resampling = FALSE)
head(sigs1)
sigs2 <- read.edf.signals(file, resampling = TRUE, f.new = 128)
head(sigs2)

}
