% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.empi.db.file.R
\name{read.empi.db.file}
\alias{read.empi.db.file}
\title{Reads data from a SQLite file created by the Matching Pursuit algorithm}
\usage{
read.empi.db.file(db.file)
}
\arguments{
\item{db.file}{SQLite file.}
}
\value{
1) Detailed parameters of all the generated atoms. 2) Analysed input signal(s). 3) Reconstructed signal (as the sum of generated atoms. 4) Generated Gabor atoms.
}
\description{
Reads data from a SQLite file created by the Matching Pursuit algorithm
}
\examples{
## Not run:
file <- system.file("extdata", "EEG_data_10sec.db", package = "MatchingPursuit")
out <- read.empi.db.file(file)

n.channnels <- ncol(out$original.signal)
original.signal <- out$original.signal
reconstruction <- out$reconstruction
t <- out$t
f <- out$f

par(mfrow = c(2, 1), pty = "m")
par(mai = c(0.9, 0.5, 0.2, 0.4))

plot(
  original.signal[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n.channnels, " (original signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

len <- length(original.signal[, 1])
lab <- seq(t[1], t[len] + 1 / f, length.out = 11)
axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)

plot(
  reconstruction[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n.channnels, " (reconstructed signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)
## End(Not run)

}
