% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_signals.R
\name{read_csv_signals}
\alias{read_csv_signals}
\title{Reads and validates a CSV file structure}
\usage{
read_csv_signals(file, col_names = NULL, col_names_in_csv = FALSE)
}
\arguments{
\item{file}{File to be read and checked. The first line of the file must contain two numbers:
the sampling frequency in Hz (\code{freq}) and the signal length in seconds (\code{sec}).
The function verifies whether the file contains exactly \code{round(freq * sec)} samples.
The two numbers must be separated by one or more whitespace characters.}

\item{col_names}{Optional character vector of column names. If not specified, default names are created.}

\item{col_names_in_csv}{Logical value. If \code{TRUE}, the second line of the file is assumed
to contain column names.}
}
\value{
A list containing:

\describe{
  \item{signal}{Data frame containing all signals (rows = samples, columns = channels).}
  \item{sampling_frequency}{Sampling frequency.}
  \item{time}{Time vector corresponding to signal samples.}
}
}
\description{
Reads and validates a CSV file structure
}
\examples{
file <- system.file("extdata", "sample1.csv", package = "MatchingPursuit")

# The first line of the file must contain two numbers:
# a) the sampling frequency in Hz
# b) the signal length in seconds
out <- read.csv(file, header = FALSE)
head(out)

signal <- read_csv_signals(file, col_names = "signal_1")
head(signal$signal)
signal$sampling_frequency
head(signal$time)
tail(signal$time)

file <- system.file("extdata", "sample2.csv", package = "MatchingPursuit")
signal <- read_csv_signals(file, col_names = c("signal_1"))
head(signal$signal)
signal$sampling_frequency

# Now, the csv file contains signal names in the second line
file <- system.file("extdata", "sample3.csv", package = "MatchingPursuit")
signal <- read_csv_signals(file, col_names_in_csv = TRUE)
head(signal$signal)
signal$sampling_frequency

}
