% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edf_signals.R
\name{read_edf_signals}
\alias{read_edf_signals}
\title{Reads a selected EDF or EDF+ file and returns signal data}
\usage{
read_edf_signals(
  file,
  resampling = FALSE,
  sf_new = NULL,
  from = NULL,
  to = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file}{Path to the EDF / EDF+ file to be read.}

\item{resampling}{If \code{TRUE}, all signals are resampled
(either upsampled or downsampled), depending on the original sampling rates of the channels.}

\item{sf_new}{Target sampling frequency used for upsampling or downsampling.}

\item{from}{Starting time of the signal to be loaded (in seconds).}

\item{to}{Ending time of the signal to be loaded (in seconds).}

\item{verbose}{Logical flag indicating whether progress information should be printed.}
}
\value{
An object of class \code{edf}, which is a list with fields:

\item{signal}{Data frame containing all signal channels.}
\item{sampling_frequency}{Sampling frequency after optional resampling.}
\item{time_stamps}{Time stamps after optional resampling.}
\item{signal_names}{Names of the signal channels.}
\item{record_name}{Name of the EDF file.}
}
\description{
The function reads a selected EDF or EDF+ file. Optionally, resampling can be performed
(upsampling or downsampling).
}
\details{
If \code{resampling = TRUE}, signals are resampled according to the target frequency
specified by \code{f.new}. Since the EDF standard allows different sampling rates per channel,
some channels may be upsampled while others are downsampled. The function does not support
independent resampling of individual channels.
}
\examples{
# Read EDF signals without resampling
file <- system.file("extdata", "EEG.edf", package = "MatchingPursuit")
out1  <- read_edf_signals(file, resampling = FALSE)

lapply(out1, class)
out1$sampling_frequency

# Read EDF signals and resample them to 128 Hz
out2 <- read_edf_signals(file, resampling = TRUE, sf_new = 128, verbose = TRUE)

lapply(out2, class)
out2$sampling_frequency

}
