% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_empi_db_file.R
\name{read_empi_db_file}
\alias{read_empi_db_file}
\title{Reads data from a SQLite file created by the Matching Pursuit algorithm}
\usage{
read_empi_db_file(db_file)
}
\arguments{
\item{db_file}{SQLite file.}
}
\value{
An object of class \code{"mp"} containing:

\item{atoms}{A data frame describing the selected atoms.}
\item{original_signal}{Matrix containing the original signal(s).}
\item{reconstruction}{Matrix containing the reconstructed signal(s).}
\item{gabors}{List of matrices containing selected atoms for each channel.}
\item{t}{Time vector corresponding to signal samples.}
\item{sf}{Sampling frequency.}
}
\description{
Reads data from a SQLite file (\code{.db}) created by the Matching Pursuit algorithm.
The reconstructed signal(s) and Gabor function(s) are also returned.
}
\examples{
file <- system.file("extdata", "EEG_bipolar_filtered.db", package = "MatchingPursuit")
out <- read_empi_db_file(file)

n_channels <- ncol(out$original_signal)
original_signal <- out$original_signal
reconstruction <- out$reconstruction
t <- out$t
sf <- out$sf

old.par <- par("mfrow", "pty", "mai")

par(mfrow = c(2, 1))
par(pty = "m")
par(mai = c(0.9, 0.5, 0.3, 0.4))

plot(
  original_signal[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n_channels, " (original signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

len <- length(original_signal[, 1])
lab <- seq(t[1], t[len] + 1 / sf, length.out = 11)
axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)

plot(
  reconstruction[,1], type = "l", col = "blue",
  main = paste("channel: ", 1, " / " , n_channels, " (reconstructed signal)",  sep = ""),
  xaxt = "n", ylab = "", xlab = "time [sec]"
)

axis(side = 1, las = 1, cex.axis = 0.9, at = seq(0, len, length.out = 11), labels = lab)

par(old.par)

}
