% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2bin.R
\name{sig2bin}
\alias{sig2bin}
\title{Reads input signal(s) from a data frame and returns them in binary format}
\usage{
sig2bin(data, write_to_file = FALSE, path = NULL, file_name = NULL)
}
\arguments{
\item{data}{Data frame containing the input signal(s).}

\item{write_to_file}{If \code{TRUE}, a \code{.bin} file is created and saved in the
\code{path} directory or, if \code{path = NULL}, in the cache directory.}

\item{path}{Directory in which the binary file will be saved.
If \code{NULL}, the file will be saved in the cache directory.}

\item{file_name}{Name of the file to create if \code{write_to_file = TRUE}.}
}
\value{
A \code{raw} vector containing the binary representation of the signal.
If \code{write_to_file = TRUE}, a \code{.bin} file is additionally created.
}
\description{
Saves the given data (signals) in binary form. The input signal(s) must be a data frame:
rows correspond to samples for all channels, and columns correspond to channels.
The function is used internally by \code{empi_execute()}. The binary data consist of
4-byte floating-point values. Values are written using little-endian byte order.

For multichannel signals, the binary output is written in time order: first all
channels at \code{t = 0}, then all channels at at \code{t = }\eqn{\Delta t},
and so on. In other words, the signal is stored in column-major order
(rows = channels, columns = samples).
}
\note{
The \code{.bin} files generated by this function are not intended for direct
user manipulation. They are used internally by \code{empi_execute()}. The external
program \emph{Enhanced Matching Pursuit Implementation} (EMPI) requires binary input
data. This conversion utility may also be useful for users who wish to run EMPI
outside of the R environment.
}
\examples{
file <- system.file("extdata", "sample3.csv", package = "MatchingPursuit")
out <- read_csv_signals(file, col_names_in_csv = TRUE)

signal_bin <- sig2bin(data = out$signal, write_to_file = FALSE)

# We have 3 channels. The first 4 time points.
head(out$signal, 4)

# The same elements of the signal in binary (floats are stored in 4 bytes).
head(signal_bin, 48)

# After decoding to numeric.
# Of course we get the same values as in out$signal.
readBin(signal_bin[1:4], what = "numeric", size = 4, endian = "little")
readBin(signal_bin[5:8], what = "numeric", size = 4, endian = "little")
readBin(signal_bin[41:44], what = "numeric", size = 4, endian = "little")
readBin(signal_bin[45:48], what = "numeric", size = 4, endian = "little")

}
