% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_map.R
\name{tf_map}
\alias{tf_map}
\title{Creates a time-frequency map using atoms from the Matching Pursuit algorithm}
\usage{
tf_map(
  x = NULL,
  channel,
  mode = "sqrt",
  freq_divide = NULL,
  increase_factor = 1,
  shortening_factor_x = 2,
  shortening_factor_y = 2,
  display_crosses = TRUE,
  display_atom_numbers = FALSE,
  display_grid = FALSE,
  color = "white",
  palette = "my custom palette",
  reverse_palette = TRUE,
  out_mode = "plot",
  path = NULL,
  file_name = NULL,
  size = c(512, 512),
  draw_ellipses = FALSE,
  plot_signals = TRUE,
  write_atoms = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{An object of class \code{mp} or a path to a SQLite file created by \code{empi_execute()}.}

\item{channel}{Channel from the SQLite file to process.}

\item{mode}{\code{"sqrt"}, \code{"log"}, or \code{"linear"}. Determines the intensity
with which the so-called blobs are displayed on the T-F map.}

\item{freq_divide}{Specifies how many times the displayed frequency range in the T-F map
should be reduced. At high sampling rates, especially when a low-pass filter with
a cut-off frequency much lower than the sampling frequency is used, a large part of
the T-F map may contain no blobs. If the sampling frequency is \code{f},
the maximum frequency displayed in the T-F map will be \code{ceiling(f / 2 / freq_divide)}
(\code{f / 2} follows the Nyquist rule). If \code{NULL}, it is determined from the atom
with the highest frequency \code{fmax} according to \code{freq_divide = (f / 2) / fmax}.}

\item{increase_factor}{Factor controlling the increase in the number of pixels along the
frequency axis. Non-negative integers such as 2, 4, 5, or 8 are usually appropriate.}

\item{shortening_factor_x}{Usually, a value of 2 provides better atom visualization.}

\item{shortening_factor_y}{Usually, a value of 2 provides better atom visualization.}

\item{display_crosses}{Whether small crosses should be displayed at the centres of atoms.}

\item{display_atom_numbers}{Whether atom numbers should be displayed in the cantres of atoms.}

\item{display_grid}{Whether grid lines should be drawn.}

\item{color}{Color of the small crosses or atom numbers.}

\item{palette}{Palette from the list returned by the \code{hcl.pals()} function or the string
\code{"my custom palette"}.}

\item{reverse_palette}{Value of the \code{rev} argument passed to the \code{hcl.colors()} function.}

\item{out_mode}{One of the following:
\itemize{
   \item \code{"plot"} - draws a T-F map on the screen.
   \item \code{"file"} - saves a T-F map to the file \code{file_name} (as a \code{png} file).
   \item \code{"RData"} - saves the T-F map of size \code{size} to \code{file_name}
   (as an R matrix); resampling is performed using the \code{imager::resize()} function.
   \item \code{"RData2"} - saves the T-F map of size \code{size} to \code{file_name}
   (as an R matrix); resampling is performed using the \code{raster::resample()} function.
 }}

\item{path}{Path where \code{png}, \code{RData}, or \code{pdf} files will be written. If \code{NULL},
files will be written to the cache directory.}

\item{file_name}{Name of the \code{png} file (if \code{out_mode = "file"}) or name of the
\code{RData} file (if \code{out_mode = "RData"} or \code{out_mode = "RData2"}).}

\item{size}{Size of the \code{png} file in pixels (if \code{out_mode = "file"}) or size of
the T-F matrix (if \code{out_mode = "RData"} or \code{out_mode = "RData2"}).}

\item{draw_ellipses}{Intended for testing only. Can be set to \code{TRUE} to display
the effect. Works correctly only if \code{out_mode = "plot"}.}

\item{plot_signals}{Whether the original and reconstructed signals should also be displayed.}

\item{write_atoms}{If \code{TRUE}, writes all atom plots to the \code{Atoms.pdf} file
(in the cache directory or in a user-specified directory, depending on \code{path}).}

\item{verbose}{Logical flag indicating whether progress information should be printed.}
}
\value{
Depending on the \code{out_mode} parameter, the function:
   \itemize{
     \item displays the time-frequency map on the screen
     \item saves the time-frequency map as a \code{.png} file
     \item saves the time-frequency map as a \code{.RData} file
  }
Regardless of the output mode, the function also returns:

    \item{gabor_functions}{All Gabor functions.}
    \item{reconstruction}{Reconstructed signal.}
    \item{signal}{ Original signal.}
    \item{sampling_frequency}{Sampling frequency.}
    \item{grid_size_t}{Grid size along the time axis.}
    \item{grid_size_f}{Grid size along the frequency axis.}
    \item{epochSize}{Epoch size in samples.}
    \item{number_of_secs}{Signal length in seconds.}
    \item{tf_map}{Time-frequency map.}
    \item{tf_map_resampled}{Resampled time-frequency map
    (if \code{out_mode = "RData"} or \code{out_mode = "RData2"}; otherwise \code{NULL}).}
    \item{channel}{Processed channel number.}
    \item{freq_divide}{Frequency division factor.}
}
\description{
Creates a time-frequency map using atoms from the Matching Pursuit algorithm.
The resulting map can be: 1) displayed on the screen, 2) saved as a \code{.png} file,
or 3) saved as an \code{.RData} object.
}
\examples{
file <- system.file("extdata", "sample1.db", package = "MatchingPursuit")
empi_class <- read_empi_db_file(file)

# 'freq_divide' is set arbitrarily
out <- tf_map(
  x = empi_class,
  channel = 1,
  mode = "sqrt",
  freq_divide = 4,
  increase_factor= 4,
  display_crosses = TRUE,
  display_atom_numbers = FALSE,
  out_mode = "plot",
)

# 'freq_divide' is determined based on the atom with the highest frequency
out <- tf_map(
  x = empi_class,
  channel = 1,
  mode = "sqrt",
  increase_factor= 4,
  display_crosses = TRUE,
  display_atom_numbers = FALSE,
  out_mode = "plot",
)

}
