% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topk-methods.R
\name{topk-methods}
\alias{topk-methods}
\alias{print.topk}
\alias{summary.topk}
\alias{print.summary.topk}
\title{Methods for objects of class "topk"}
\usage{
\method{print}{topk}(x, ...)

\method{summary}{topk}(object, ...)

\method{print}{summary.topk}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.topk"}.}

\item{...}{Additional arguments, currently ignored.}

\item{object}{An object of class \code{"topk"}.}
}
\value{
\code{print.topk()} returns \code{x} invisibly.

\code{summary.topk()} returns an object of class \code{"summary.topk"}
containing basic information about the selected atoms and their parameters.

\code{print.summary.topk()} returns \code{x} invisibly.
}
\description{
Print and summarize objects returned by \code{topk_gabor_atoms()}.
}
\seealso{
\code{\link{topk_gabor_atoms}}
}
